### 2024-03-06 #164 @kyleaoman
Martini
Submitting Author: Kyle Oman (@kyleaoman)
Package Name: martini
One-Line Description of Package:  MARTINI is a modular package for the creation of synthetic resolved HI line observations (data cubes) of smoothed-particle hydrodynamics simulations of galaxies.
Repository Link:  https://github.com/kyleaoman/martini
Version submitted: 2.0.11 (note JOSS paper is in branch joss-paper, and that branch is somewhat behind main & 2.0.11)
Archive: https://zenodo.org/doi/10.5281/zenodo.11193206
JOSS DOI: https://doi.org/10.21105/joss.06860
Version accepted: 2.0.15

---

MARTINI is a modular package for the creation of synthetic resolved HI line observations (data cubes) of smoothed-particle hydrodynamics simulations of galaxies. The various aspects of the mock-observing process are divided logically into sub-modules handling the data cube, source, beam, noise, spectral model and SPH kernel. MARTINI is object-oriented: each sub-module provides a class (or classes) which can be configured as desired. For most sub-modules, base classes are provided to allow for straightforward customization. Instances of each sub-module class are given as parameters to the Martini class; a mock observation is then constructed by calling a handful of functions to execute the desired steps in the mock-observing process.

- [x] Astropy: Link coming soon to standards

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience is research astronomers interested in galaxies (broadly, "extragalactic astronomers") from both the theoretical and observational communities. The package provides a way to transform data products from the theory community (smoothed-particle hydrodynamics based simulations of galaxy formation and evolution) into data products closely resembling the atomic hydrogen signal observed with a radio telescope at 21-cm wavelengths. This enables much more faithful comparisons between theoretical predictions and measurements.

I am not aware of any other actively maintained packages with similar purpose.

N/A (I had previously submitted to astropy under their affiliated package scheme and have been redirected here)

- [x] The package is deposited in a long-term repository with the DOI: https://zenodo.org/doi/10.5281/zenodo.11193206

Since I'm coming in through the new astropy route, I had not prepared for JOSS requirements, but was thinking of submitting to JOSS soon anyway. I've now pushed a `paper.md` in the `joss-paper` branch. My package is indexed in the [ASCL](https://ascl.net/1911.005) which in turn is indexed in [ADS](https://ui.adsabs.harvard.edu/abs/2019ascl.soft11005O/abstract), but neither of these seems to provide an actual DOI so I will need to look into other repositories (probably Zenodo). As far as I understand submission to JOSS happens after the pyOpenSci review so there is a little bit of time to get this done - I expect to have these two items ticked off by the time the pyOpenSci review process is completed.
