### 2024-03-12 215af61
Initialize paper skeleton.
