+++ paper.md
@@ -1,0 +1,39 @@
+---
+title: "MARTINI: Mock Array Radio Telescope Interferometry of the Neutral ISM"
+tags:
+  - Python
+  - astronomy
+  - simulations
+authors: 
+  - name: Kyle A. Oman
+    orcid: 0000-0001-9857-7788
+    affiliation: 1,2
+affiliations:
+  - name: Institute for Computational Cosmology, Physics Department, Durham University
+    index: 1
+  - name: Centre for Extragalactic Astronomy, Physics Department, Durham University
+    index: 2
+date: 13 March 2024
+codeRepository: https://github.com/kyleaoman/martini
+license: LGPLv3
+bibliography: bibliography.bib
+---
+
+# Summary
+
+
+# Background
+
+
+# Why MARTINI?
+
+
+`MARTINI` is hosted on [GitHub](https://github.com/kyleaoman/martini) and
+has documentation available through
+[ReadTheDocs](https://martini.readthedocs.io).
+
+# Acknowledgements
+
+KAO acknowledges support by the Royal Society trhough Dorothy Hodgkin Fellowship DHF/R1/231105, by STFC through grant ST/T000244/1, by the European Research Council (ERC) through an Advanced Investigator Grant to C. S. Frenk, DMIDAS (GA 786910), and by the Netherlands Foundation for Scientific Research (NWO) through VICI grant 016.130.338 to M. Verheijen. This work has made use of NASA's Astrophysics Data System.
+
+# References
