### 2024-03-13 28bb4a1
Start drafting a few paragraphs of JOSS paper.
