+++ paper.md
@@ -26,17 +26,17 @@ MARTINI is a modular Python package that takes smoothed-particle hydrodynamics (
 
 # Background
 
-The primordial Universe contained predominantly hydrogen, some helium, and trace amount of heavier elements. Hydrogen remains the most abundant element, occuring ubiquitously in stars and interstellar/intergalactic gas. Atomic hydrogen (i.e. with an electron bound to the nucleus) is much more abundant in galaxies where the gravitational field of dark matter allows gas to collect and cool. Hydrogen atoms in their lowest energy state exhibit a transition between the state where the electron spin is aligned and that where it is anti-aligned with the nuclear spin. The very close ("hyperfine") spacing in energy between these two states means that the photon emitted/absorbed by a transition between the states has a correspondingly low energy, or equivalently, long wavelength of about 21 cm. The decay from the excited (spins aligned) to the ground (spins anti-aligned) state is very slow, with a mean lifetime of about 11 million years, but the ubiquity of hydrogen in the Universe makes the 21 cm, or "HI", line readily observable in emission (and absorption, but this is not the focus of MARTINI) using radio telescopes. The fact that the 21 cm radiation originates from a spectral line means that the precise frequency in the rest frame is known from laboratory measurements, so the observed frequency can be used to measure a Doppler shift and therefore a measure of the kinematics of the emitting gas. These properties lead to a wealth of scientific applications of HI line emission observations, including in extragalactic astrophysics where MARTINI focuses. An more detailed overview of 21 cm radio astronomy is available in [cite].
+The primordial Universe contained predominantly hydrogen, some helium, and trace amount of heavier elements. Hydrogen remains the most abundant element, occuring ubiquitously in stars and interstellar/intergalactic gas. Atomic hydrogen (i.e. with an electron bound to the nucleus) is much more abundant in galaxies where the gravitational field of dark matter allows gas to collect and cool. Hydrogen atoms in their lowest energy state exhibit a transition between the state where the electron spin is aligned and that where it is anti-aligned with the nuclear spin. The very close ("hyperfine") spacing in energy between these two states means that the photon emitted/absorbed by a transition between the states has a correspondingly low energy, or equivalently, long wavelength of about 21 cm. The decay from the excited (spins aligned) to the ground (spins anti-aligned) state is very slow, with a mean lifetime of about 11 million years, but the ubiquity of hydrogen in the Universe makes the 21 cm, or "HI", line readily observable in emission (and absorption, but this is not the focus of MARTINI) using radio telescopes. The fact that the 21 cm radiation originates from a spectral line means that the precise frequency in the rest frame is known from laboratory measurements, so the observed frequency can be used to measure a Doppler shift and therefore a measure of the kinematics of the emitting gas. These properties lead to a wealth of scientific applications of HI line emission observations, including in extragalactic astrophysics where MARTINI focuses. A more detailed, forward-looking overview of the types of applications in extragalactic HI radio astronomy that MARTINI is suited to can be found in @deBlok2015.
 
[…]
