+++ bibliography.bib
@@ -1,0 +1,336 @@
+@ARTICLE{Schaye2015,
+       author = {{Schaye}, Joop and {Crain}, Robert A. and {Bower}, Richard G. and {Furlong}, Michelle and {Schaller}, Matthieu and {Theuns}, Tom and {Dalla Vecchia}, Claudio and {Frenk}, Carlos S. and {McCarthy}, I.~G. and {Helly}, John C. and {Jenkins}, Adrian and {Rosas-Guevara}, Y.~M. and {White}, Simon D.~M. and {Baes}, Maarten and {Booth}, C.~M. and {Camps}, Peter and {Navarro}, Julio F. and {Qu}, Yan and {Rahmati}, Alireza and {Sawala}, Till and {Thomas}, Peter A. and {Trayford}, James},
+        title = "{The EAGLE project: simulating the evolution and assembly of galaxies and their environments}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {methods: numerical, galaxies: evolution, galaxies: formation, cosmology: theory, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2015,
+        month = jan,
+       volume = {446},
+       number = {1},
+        pages = {521-554},
+          doi = {10.1093/mnras/stu2058},
+archivePrefix = {arXiv},
+       eprint = {1407.7040},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015MNRAS.446..521S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Vogelsberger2014,
+       author = {{Vogelsberger}, Mark and {Genel}, Shy and {Springel}, Volker and {Torrey}, Paul and {Sijacki}, Debora and {Xu}, Dandan and {Snyder}, Greg and {Nelson}, Dylan and {Hernquist}, Lars},
+        title = "{Introducing the Illustris Project: simulating the coevolution of dark and visible matter in the Universe}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {methods: numerical, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2014,
+        month = oct,
+       volume = {444},
+       number = {2},
+        pages = {1518-1547},
+          doi = {10.1093/mnras/stu1536},
+archivePrefix = {arXiv},
+       eprint = {1405.2921},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2014MNRAS.444.1518V},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Lucy1977,
+       author = {{Lucy}, L.~B.},
+        title = "{A numerical approach to the testing of the fission hypothesis.}",
+      journal = {The Astronomical Journal},
+     keywords = {Binary Stars, Hypotheses, Nuclear Fission, Numerical Analysis, Protostars, Stellar Evolution, Astronomical Models, Difference Equations, Gas Dynamics, Monte Carlo Method, Numerical Stability, Particle Motion, Stellar Rotation, Astrophysics},
+         year = 1977,
+        month = dec,
+       volume = {82},
+        pages = {1013-1024},
+          doi = {10.1086/112164},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1977AJ.....82.1013L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
[…]
