+++ paper.md
@@ -34,7 +34,7 @@ Another powerful tool for the study of galaxies, this time coming from the theor
 
 A SPH simulation of galaxies and radio astronomical observations of galaxies produce qualitatively different data sets. The output of a simulation consists of tabulated lists of particles positions, velocities, masses, chemical abundances, etc. A common high-level data product from a radio observatory is a cubic data structure where the axes of the cube correspond to right ascension (longitude on the sky), declination (latitude) and frequency. Each cell in such a cube records the intensity of radio emission with corresponding right ascension, declination and frequency. Since two of the cube axes are spatial coordinates and the third is a spectral coordinate, one can equivalently think of these data cubes as images where each pixel contains a discretized spectrum at that position instead of a single-valued intensity. Testing the theoretical predictions of simulations against these kinds of observations is therefore challenging because the data are organized into fundamentally different structures. MARTINI provides a tool to mimick the process of observing a simulated galaxy with a radio telescope, producing as output the same kinds of "data cubes" and thus enabling much more robust tests of the theoretical predictions made by galaxy formation simulations.
 
-The first cosmological hydrodynamical simulations that produced broadly realistic populations of galaxies appeared around 2014 [@Vogelsberger2014; @Schaye2015]. A few studies began to use mock observations of HI emission from these simulated galaxies shortly after this [@Read2016; @Bahe2016], including the first work using the code that would later become MARTINI [@Oman2019]. MARTINI has since been adopted by the wider research community working on these types of studies and is now used in the majority of such work including @Chauhan2019, @ManceraPina2019, @SantosSantos2020, @ManceraPina2020, @Glowacki2021, @Glowacki2022, @Bilimogga2022, @Roper2023, @Oman2024, and many more projects in progress. Other codes that build on MARTINI have also started to appear, with [ALMASim](https://github.com/MicheleDelliVeneri/ALMASim) [@Guglielmetti2023] as a first example. While some authors have developed codes that implement some of the same features as MARTINI [@Read2016], I am not aware of any such codes that have been publicly released.
+The first cosmological hydrodynamical simulations that produced broadly realistic populations of galaxies appeared around 2014 [@Vogelsberger2014; @Schaye2015]. A small number of studies began to use mock observations of HI emission from these simulated galaxies shortly after this [@Read2016], including the first work using the code that would later become MARTINI [@Oman2019]. MARTINI has since been adopted by the wider research community working on these types of studies and is now used in the majority of such work including @Chauhan2019, @ManceraPina2019, @SantosSantos2020, @ManceraPina2020, @Glowacki2021, @Glowacki2022, @Bilimogga2022, @Roper2023, @Oman2024, and many more projects in progress. Other codes that build on MARTINI have also started to appear, with [ALMASim](https://github.com/MicheleDelliVeneri/ALMASim) [@Guglielmetti2023] as a first example. While some authors have developed codes that implement some of the same features as MARTINI [@Read2016], I am not aware of any such codes that have been publicly released.
 
 MARTINI is hosted on [GitHub](https://github.com/kyleaoman/martini) and has documentation available through [ReadTheDocs](https://martini.readthedocs.io).
 
