### 2024-03-14 bcd2f19
Minor edits completing first draft.
