+++ README.rst
@@ -66,6 +66,22 @@ If your use of MARTINI leads to a publication, please acknowledge this and link 
 .. _indexed on ADS: https://ui.adsabs.harvard.edu/abs/2019ascl.soft11005O/abstract
 .. _original paper: https://ui.adsabs.harvard.edu/abs/2019MNRAS.482..821O/abstract
 
+Work that has used MARTINI includes: `Oman et al. (2019)`_, `Mancera Piña et al. (2019)`_, `Chauhan et al. (2019)`_, `Mancera Piña et al. (2020)`_, `Santos-Santos et al. (2020)`_, `Glowacki et al. (2021)`_, `Bilimogga et al. (2022)`_, `Glowacki et al. (2022)`_, `Roper et al. (2023)`_ and `Oman et al. (2024)`_. The ALMASim_ package (`Guglielmetti et al. 2023`_) builds on some of MARTINI's functionality. If your work has used MARTINI and is not listed here, please let me know (by email_ or github issue).
+
+.. _Oman et al. (2019): https://doi.org/10.1093/mnras/sty2687
+.. _Mancera Piña et al. (2019): https://doi.org/10.3847/2041-8213/ab40c7
+.. _Chauhan et al. (2019): https://doi.org/10.1093/mnras/stz2069
+.. _Mancera Piña et al. (2020): https://doi.org/10.1093/mnras/staa1256
+.. _Santos-Santos et al. (2020): https://doi.org/10.1093/mnras/staa1072
+.. _Glowacki et al. (2021): https://doi.org/10.1093/mnras/stab2279
+.. _Bilimogga et al. (2022): https://doi.org/10.1093/mnras/stac1213
+.. _Glowacki et al. (2022): https://doi.org/10.1093/mnras/stac2684
+.. _Roper et al. (2023): https://doi.org/10.1093/mnras/stad549
+.. _Oman et al. (2024): https://doi.org/10.48550/arXiv.2401.11878
+.. _ALMASim: https://github.com/MicheleDelliVeneri/ALMASim
+.. _Guglielmetti et al. 2023: https://doi.org/10.48550/arXiv.2311.10657
+.. _email: mailto:kyle.a.oman@durham.ac.uk
+
 .. CITING_END_LABEL
 
 Installation Notes
@@ -75,12 +91,12 @@ Installation Notes
 
 MARTINI works with ``python3`` (version ``3.7`` or higher), and does not support ``python2``.
 
-Stable releases are available via PyPI_ (``pip install astromartini``) and the numbered releases (starting from 2.0.0) on github. The github main branch is actively developed: things will change, bugs will happen. Any feedback is greatly appreciated via github or kyle.a.oman@durham.ac.uk.
+Stable releases are available via PyPI_ (``pip install astromartini``) and the numbered releases (starting from 2.0.0) on github. The github main branch is actively developed: things will change, bugs will happen. Any feedback is greatly appreciated via github issues or kyle.a.oman@durham.ac.uk.
 
 .. _PyPI: https://pypi.org/
 .. _kyle.a.oman@durham.ac.uk: mailto:kyle.a.oman@durham.ac.uk
 
-The easiest way to install martini is from PyPI by doing ``python3 -m pip install astromartini``. Output to ``.fits`` files is supported by default; if output to ``.hdf5`` format is desired use ``python3 -m pip install astromartini[hdf5_output]`` instead. This will also handle the installation of the required dependencies. Other optional features require additional dependencies hosted on PyPI. In particular, EAGLE, Illustris/TNG and Simba users who wish to use the custom source modules for those simulations in MARTINI can automatically install the optional dependencies with ``python3 -m pip install astromartini[eaglesource]``, or ``[simbasource]`` or ``[tngsource]``.
[…]
