### 2024-03-14 4ed3212
Explicitly check that padding is sufficient before convolving beam.
