### 2024-03-14 7abe9cf
Add docstrings for spec_dtype.
