### 2024-03-15 0ec44bb
Add a test that parallel spectral match serial spectra.
