+++ tests/test_spectral_models.py
@@ -157,13 +157,19 @@ class TestSpectrumPrecision:
         assert spectral_model.spectra.dtype == dtype
 
 
-class TestParallel:
+class TestParallelSpectra:
     @pytest.mark.parametrize("SpectralModel", spectral_models)
-    def test_parallel(self, SpectralModel, cross_source):
+    def test_parallel_spectra(self, SpectralModel, cross_source):
+        """
+        Check that spectra calculated in serial and parallel are consistent.
+        """
         source = cross_source()
-        spectral_model_serial = SpectralModel(ncpu=1)
-        spectral_model_parallel = SpectralModel(ncpu=2)
+        source._init_skycoords()
+        spectral_model_serial = SpectralModel()
+        spectral_model_parallel = SpectralModel()
         datacube = DataCube()
-        spectral_model_serial.init_spectra(source, datacube)
-        spectral_model_parallel.init_spectra(source, datacube)
-        assert U.allclose(spectral_model_serial.spectra, spectral_model_parallel.spectra)
+        spectral_model_serial.init_spectra(source, datacube, ncpu=1)
+        spectral_model_parallel.init_spectra(source, datacube, ncpu=4)
+        assert U.allclose(
+            spectral_model_serial.spectra, spectral_model_parallel.spectra
+        )
