### 2024-03-16 ae681d4
Fix some typos in kernel docs.
