### 2024-03-16 e454b4d
Fix some typos in kernel docs.
