### 2024-03-16 ad6fa4e
Better formatting of math
