### 2024-03-16 e714ef4
Rough implementation that achieves parallel speedup.
