+++ martini/spectral_models.py
@@ -27,6 +27,8 @@ class _BaseSpectrum(metaclass=ABCMeta):
 
     Parameters
     ----------
+    ncpu: int, optional
+        Number of cpus to use for evaluation of particle spectra. (Default: 1)
     spec_dtype: type, optional
         Data type of the arrays storing spectra of each particle, can be used to manage
         memory usage by adjusting precision.
@@ -38,7 +40,8 @@ class _BaseSpectrum(metaclass=ABCMeta):
     DiracDeltaSpectrum
     """
 
-    def __init__(self, spec_dtype=np.float64):
+    def __init__(self, ncpu=1, spec_dtype=np.float64):
+        self.ncpu = ncpu
         self.spectral_function_extra_data = None
         self.spectra = None
         self.spec_dtype = spec_dtype
@@ -65,9 +68,9 @@ class _BaseSpectrum(metaclass=ABCMeta):
             spectral channels.
         """
 
-        channel_edges = datacube.channel_edges
-        channel_widths = np.diff(channel_edges).to(U.km * U.s**-1)
-        vmids = source.skycoords.radial_velocity
+        self.channel_edges = datacube.channel_edges
+        channel_widths = np.diff(self.channel_edges).to(U.km * U.s**-1)
+        self.vmids = source.skycoords.radial_velocity
         A = source.mHI_g * np.power(source.skycoords.distance.to(U.Mpc), -2)
         MHI_Jy = (
             U.Msun * U.Mpc**-2 * (U.km * U.s**-1) ** -1,
@@ -75,30 +78,20 @@ class _BaseSpectrum(metaclass=ABCMeta):
             lambda x: (1 / 2.36e5) * x,
             lambda x: 2.36e5 * x,
         )
-        self.init_spectral_function_extra_data(source, datacube)
-        channel_edges_unit = channel_edges.unit
-        vmids_unit = vmids.unit
-        raw_spectra = self.spectral_function(
-            (
-                np.tile(
-                    channel_edges.to_value(channel_edges_unit)[:-1], vmids.shape + (1,)
+        if self.ncpu == 1:
+            raw_spectra = self.evaluate_spectra(source, datacube)
+        else:
+            from multiprocess.pool import Pool
+
+            with Pool(processes=self.ncpu) as pool:
+                raw_spectra = np.vstack(
+                    pool.map(
+                        lambda icpu: self.evaluate_spectra(
+                            source, datacube, parallel=icpu
+                        ),
+                        range(self.ncpu),
+                    )
                 )
-                * channel_edges.unit
-            ).astype(self.spec_dtype),
-            (
-                np.tile(
-                    channel_edges.to_value(channel_edges_unit)[1:], vmids.shape + (1,)
-                )
-                * channel_edges.unit
-            ).astype(self.spec_dtype),
-            (
-                np.tile(
-                    vmids.to_value(vmids_unit),
-                    np.shape(channel_edges[:-1]) + (1,) * vmids.ndim,
-                ).T
-                * vmids.unit
-            ).astype(self.spec_dtype),
-        ).to_value(U.dimensionless_unscaled)
         self.spectra = (
             A.astype(self.spec_dtype)[..., np.newaxis]
             * raw_spectra
@@ -106,6 +99,47 @@ class _BaseSpectrum(metaclass=ABCMeta):
         ).to(U.Jy, equivalencies=[MHI_Jy])
 
         return
+
+    def evaluate_spectra(self, source, datacube, parallel=None):
+        """
+        TODO DOCSTRING
+        """
+        mask = (
+            np.s_[
+                parallel
+                * len(self.vmids)
+                // self.ncpu : (parallel + 1)
+                * len(self.vmids)
+                // self.ncpu
+            ]
+            if parallel is not None
[…]
