### 2024-03-16 62b7ad3
Tidy up docstrings, stubs.
