+++ martini/spectral_models.py
@@ -27,9 +27,11 @@ class _BaseSpectrum(metaclass=ABCMeta):
 
     Parameters
     ----------
-    ncpu: int, optional
-        Number of cpus to use for evaluation of particle spectra. (Default: 1)
-    spec_dtype: type, optional
+    ncpu : int, optional
+        Number of cpus to use for evaluation of particle spectra. Defaults to 1 if not
+        provided. (Default: None)
+
+    spec_dtype : type, optional
         Data type of the arrays storing spectra of each particle, can be used to manage
         memory usage by adjusting precision.
 
@@ -40,8 +42,8 @@ class _BaseSpectrum(metaclass=ABCMeta):
     DiracDeltaSpectrum
     """
 
-    def __init__(self, ncpu=1, spec_dtype=np.float64):
-        self.ncpu = ncpu
+    def __init__(self, ncpu=None, spec_dtype=np.float64):
+        self.ncpu = ncpu if ncpu is not None else 1
         self.spectral_function_extra_data = None
         self.spectra = None
         self.spec_dtype = spec_dtype
@@ -90,10 +92,21 @@ class _BaseSpectrum(metaclass=ABCMeta):
             with Pool(processes=self.ncpu) as pool:
                 raw_spectra = np.vstack(
                     pool.map(
-                        lambda icpu: self.evaluate_spectra(
-                            source, datacube, parallel=icpu
-                        ),
-                        range(self.ncpu),
+                        lambda mask: self.evaluate_spectra(source, datacube, mask=mask),
+                        [
+                            (
+                                np.s_[
+                                    icpu
+                                    * len(self.vmids)
+                                    // self.ncpu : (icpu + 1)
+                                    * len(self.vmids)
+                                    // self.ncpu
+                                ]
+                                if icpu is not None
+                                else np.s_[...]
+                            )
+                            for icpu in range(self.ncpu)
+                        ],
                     )
                 )
         self.spectra = (
@@ -104,7 +117,7 @@ class _BaseSpectrum(metaclass=ABCMeta):
 
         return
 
-    def evaluate_spectra(self, source, datacube, parallel=None):
+    def evaluate_spectra(self, source, datacube, mask=np.s_[...]):
         """
         The main portion of the calculation of the spectra.
 
@@ -122,24 +135,11 @@ class _BaseSpectrum(metaclass=ABCMeta):
             DataCube object defining the observational parameters, including
             spectral channels.
 
-        parallel: int, optional
-            Rank of the parallel process, used to choose a segment of the source
-            line-of-sight velocity values to operate on. If None, the calculation
-            is done in serial. (Default: None)
+        mask : slice, optional
+            Slice defining the subset of particles to operate on. (Default: np.s_[...])
         """
-        mask = (
-            np.s_[
-                parallel
-                * len(self.vmids)
-                // self.ncpu : (parallel + 1)
-                * len(self.vmids)
-                // self.ncpu
-            ]
-            if parallel is not None
-            else np.s_[...]
-        )
         vmids = self.vmids[mask]
-        self.init_spectral_function_extra_data(source, datacube, parallel=parallel)
+        self.init_spectral_function_extra_data(source, datacube, mask=mask)
         return self.spectral_function(
             (
                 np.tile(
@@ -204,7 +204,7 @@ class _BaseSpectrum(metaclass=ABCMeta):
 
         pass
 
-    def init_spectral_function_extra_data(self, source, datacube, parallel=None):
[…]
