### 2024-03-16 4a43931
Add narrative docs for spectral models.
