### 2024-03-16 d2a9049
Adjust tests for parallel spectra, bugfix for scalar linewidth.
