### 2024-03-17 e62366b
Demo source creation moved to separate function.
