### 2024-03-17 5999eb1
Add a source module method to make preview plot.
