### 2024-03-17 8653eee
Inject demo_source into base namespace.
