+++ martini/__init__.py
@@ -1,6 +1,6 @@
 from .martini import Martini
 from .datacube import DataCube
-from ._demo import demo
+from ._demo import demo, demo_source
 from .__version__ import __version__
 
 __doc__ = """
