### 2024-03-17 980dae3
Add scatter to demo hsm, mHI for preview demo.
