### 2024-03-17 83a7e9c
Refine point sizing/opacity in preview.
