+++ martini/sources/sph_source.py
@@ -351,7 +351,16 @@ class SPHSource(object):
         np.savetxt(fname, self.current_rotation)
         return
 
-    def preview(self, max_points=500, fig=1, lim=None, vlim=None, title="", save=None):
+    def preview(
+        self,
+        max_points=1000,
+        fig=1,
+        lim=None,
+        vlim=None,
+        point_scaling="auto",
+        title="",
+        save=None,
+    ):
         """
         Produce a figure showing the source particle coordinates and velocities.
 
@@ -365,7 +374,7 @@ class SPHSource(object):
         ----------
         max_points : int, optional
             Maximum number of points to draw per panel, the particles will be randomly
-            subsampled if the source has more. (Default: 500)
+            subsampled if the source has more. (Default: 1000)
 
         fig : int, optional
             Number of the figure in matplotlib, it will be created as `plt.figure(fig)`.
@@ -379,6 +388,14 @@ class SPHSource(object):
             The velocity axes and colour bar extend from -vlim to vlim. If unspecified,
             the maximum absolute velocity of particles in the source is used.
             (Default: None)
+
+        point_scaling : str, optional
+            By default points are scaled in size and transparency according to their HI
+            mass and the smoothing length (loosely proportional to their surface
+            densities, but with different scaling to achieve a visually useful plot). For
+            some sources the automatic scaling may not give useful results, using
+            point_scaling="fixed" will plot points of constant size without opacity.
+            (Default: "auto")
 
         title : str, optional
             A title for the figure can be provided. (Default: "")
@@ -408,28 +425,51 @@ class SPHSource(object):
             if vlim is None
             else vlim.to_value(U.km / U.s)
         )
+        alpha = (
+            (
+                (self.mHI_g[mask] / self.mHI_g.max()).to_value(U.dimensionless_unscaled)
+                * (
+                    1
+                    - (self.hsm_g[mask] / self.hsm_g.max()).to_value(
+                        U.dimensionless_unscaled
+                    )
+                )
+            )
+            ** 0.3
+            if point_scaling == "auto"
+            else 1.0
+        )
+        size = (
+            1e5 * (self.hsm_g[mask].to_value(U.kpc) / vlim) ** 2
+            if point_scaling == "auto"
+            else 10
+        )
         fig = plt.figure(fig, figsize=(12, 4))
         fig.suptitle(title)
+
         # ----- MOMENT 1 -----
         sp1 = fig.add_subplot(1, 3, 1, aspect="equal")
         sp1.set_xlim((-lim, lim))
         sp1.set_ylim((-lim, lim))
         scatter = sp1.scatter(
             self.coordinates_g.y[mask].to_value(U.kpc),
-            self.coordinages_g.z[mask].to_value(U.kpc),
+            self.coordinates_g.z[mask].to_value(U.kpc),
             c=self.coordinates_g.differentials["s"].d_x[mask].to_value(U.km / U.s),
             marker="o",
-            s=1,
-            alpha=0.5,
+            cmap="coolwarm",
+            edgecolor="None",
+            s=size,
+            alpha=alpha,
             vmin=-vlim,
             vmax=vlim,
             zorder=0,
         )
-        sp1.plot([0], [0], marker="+", ls="None", mfc="grey", ms=2, zorder=1)
[…]
