### 2024-03-17 e375a45
Compare hsm to lim, not vlim!
