### 2024-03-18 64dc198
Adjust point scaling/transparency to give reasonable results for EAGLE/TNG/Simba.
