### 2024-03-18 ac99834
Allow varying number of particles in demo.
