### 2024-03-18 eb0dd52
Adjust demo hsm_g for preview image.
