### 2024-03-18 d967d00
Outline and start drafting narrative docs for sources.
