### 2024-03-19 c46eac5
Fix typos in rotation description
