### 2024-03-19 fea1364
Finish drafting narrative docs for source modules.
