+++ docs/source/sources/index.rst
@@ -6,11 +6,12 @@ In the context of mock observations of a simulation, the source is the collectio
 Sources in MARTINI
 ------------------
 
-MARTINI has a collection of classes used to store and manipulate the properties of the source particles. All of these inherit from the :class:`martini.sources.SPHSource` class. This base class can be created by providing arrays containing the particle coordinates, velocities, masses, etc. The other source classes are tailored to specific simulations or data formats and will typically expect one or more filenames as input, along with an identifier specifying a group of particles (as defined in a group catalogue, for example). These classes will then take care of reading data from the input files, annotating them with units, doing any required calculations or conversions, and so on. There are source classes for the EAGLE_, IllustrisTNG, Simba, Colibre and Magneticum simulations. The SWIFTGalaxy_ interface to `SWIFT`_-based simulations with a variety of halo catalogue formats is also supported.
+MARTINI has a collection of classes used to store and manipulate the properties of the source particles. All of these inherit from the :class:`martini.sources.sph_source.SPHSource` class. This base class can be created by providing arrays containing the particle coordinates, velocities, masses, etc. The other source classes are tailored to specific simulations or data formats and will typically expect one or more filenames as input, along with an identifier specifying a group of particles (as defined in a group catalogue, for example). These classes will then take care of reading data from the input files, annotating them with units, doing any required calculations or conversions, and so on. There are source classes for the EAGLE_, IllustrisTNG_, Simba_, Colibre and Magneticum_ simulations. The SWIFTGalaxy_ interface to `SWIFT`_-based simulations with a variety of halo catalogue formats is also supported.
 
 .. _EAGLE: https://icc.dur.ac.uk/Eagle
 .. _IllustrisTNG: https://www.tng-project.org
 .. _Simba: http://simba.roe.ac.uk
+.. _Magneticum: http://www.magneticum.org/
 .. _SWIFTGalaxy: https://github.com/SWIFTSIM/swiftgalaxy
 .. _SWIFT: https://github.com/SWIFTSIM/SWIFT
 
@@ -54,27 +55,165 @@ Using MARTINI's source classes
 
 Quick overview here (including preview), details below
 
+Particle arrays
++++++++++++++++
+
+The core information required by the :class:`~martini.sources.sph_source.SPHSource` module is the set of arrays containing the properties of the particles making up the source. These are particle coordinates ``xyz_g``, velocities ``vxyz_g``, HI masses ``mHI_g``, smoothing lengths ``hsm_g`` and temperatures ``T_g``. All of these require units attached (see above).
+
[…]
