### 2024-03-19 9978c01
Document accessing source arrays.
