### 2024-03-19 d2ff3fb
Make preview safe with scalar input (e.g. hsm_g).
