### 2024-03-19 09de942
Tidy up formatting in martini narrative doc page.
