+++ docs/source/martini/index.rst
@@ -26,7 +26,7 @@ Once all of the component modules are set up, creating an instance of :class:`~m
 	spectral_model=spectral_model,
     )
 
-The arguments to the various modules are omitted here (replaced with `...`), check the documentation pages of each module for details. The `source`, `datacube`, `sph_kernel` and `spectral_model` arguments are mandatory. The `beam` is optional in case you want an "intrinsic" observation of the source without convolution with a beam, and the `noise` is also optional in case you don't want any in your mock observation (or perhaps want to later insert your mock into an observed noise cube). There is one more optional argument `quiet` (defaulting to `False`) that can be switched on for batch jobs where you don't want any log messages.
+The arguments to the various modules are omitted here (replaced with ``...``), check the documentation pages of each module for details. The ``source``, ``datacube``, ``sph_kernel`` and ``spectral_model`` arguments are mandatory. The ``beam`` is optional in case you want an "intrinsic" observation of the source without convolution with a beam, and the ``noise`` is also optional in case you don't want any in your mock observation (or perhaps want to later insert your mock into an observed noise cube). There is one more optional argument ``quiet`` (defaulting to ``False``) that can be switched on for batch jobs where you don't want any log messages.
 
 A few things happen behind the scenes when the :class:`~martini.martini.Martini` object is initialized:
 
@@ -44,7 +44,7 @@ This is the crucial step in creating a mock observation - the flux from the simu
 
     m.insert_source_in_cube()
 
-Since this is the most computationally demanding step in MARTINI, a progress bar is displayed by default. This can be suppressed by passing the argument `progressbar=False` (or enabled with `progressbar=True` if :class:`~martini.martini.Martini` was initialized with `quiet=True`). There is another optional argument `skip_validation`. Setting this to `True` disables internal accuracy checks and is only intended for experimentation/prototyping and code development; it should never be used for science (and anyway doesn't have any benefit in terms of e.g. speed).
+Since this is the most computationally demanding step in MARTINI, a progress bar is displayed by default. This can be suppressed by passing the argument ``progressbar=False`` (or enabled with ``progressbar=True`` if :class:`~martini.martini.Martini` was initialized with ``quiet=True``). There is another optional argument ``skip_validation``. Setting this to ``True`` disables internal accuracy checks and is only intended for experimentation/prototyping and code development; it should never be used for science (and anyway doesn't have any benefit in terms of e.g. speed).
 
 Parallelization
 +++++++++++++++
@@ -53,7 +53,7 @@ Parallelization
 
    Available since `v2.0.4`.
 
-The core loop in the source insertion function is embarassingly parallel. Parallel execution is implemented using the `multiprocess`_ package. You may need to install this, for instance `pip install multiprocess` to install from PyPI. To make use of the parallelization simply specify the number of processes to use, for example:
+The core loop in the source insertion function is embarassingly parallel. Parallel execution is implemented using the `multiprocess`_ package. You may need to install this, for instance ``pip install multiprocess`` to install from PyPI. To make use of the parallelization simply specify the number of processes to use, for example:
 
 .. _multiprocess: https://pypi.org/project/multiprocess/
 
@@ -67,12 +67,12 @@ Progress bars work in principle in parallel mode, with one bar per process, alth
 
 .. warning::
 
[…]
