### 2024-03-22 38b6c42
Add a function to Martini class to preview source particles and datacube region.
