+++ martini/martini.py
@@ -896,3 +896,233 @@ class Martini:
         if self.beam is not None:
             self.datacube.add_pad(self.beam.needs_pad())
         return
+
+    def preview(
+        self,
+        max_points=5000,
+        fig=1,
+        lim=None,
+        vlim=None,
+        point_scaling="auto",
+        title="",
+        save=None,
+    ):
+        """
+        Produce a figure showing the source particle coordinates and velocities, and the
+        datacube region.
+
+        Makes a 3-panel figure showing the projection of the source as it will appear in
+        the mock observation. The first panel shows the particles in the y-z plane,
+        coloured by the x-component of velocity (MARTINI projects the source along the
+        x-axis). The second and third panels are position-velocity diagrams showing the
+        x-component of velocity against the y and z coordinates, respectively. The red
+        boxes drawn on the panels show the datacube extent in position and velocity at the
+        distance and systemic velocity of the source.
+
+        Parameters
+        ----------
+        max_points : int, optional
+            Maximum number of points to draw per panel, the particles will be randomly
+            subsampled if the source has more. (Default: 1000)
+
+        fig : int, optional
+            Number of the figure in matplotlib, it will be created as `plt.figure(fig)`.
+            (Default: 1)
+
+        lim : Quantity with dimensions of length, optional
+            The coordinate axes extend from -lim to lim. If unspecified, the maximum
+            absolute coordinate of particles in the source is used. Whether specified
+            or not, the axes are expanded if necessary to contain the extent of the
+            data cube. Alternatively, the string "datacube" can be passed. In this case
+            the axis limits are set by the extent of the data cube. (Default: None)
+
+        vlim : Quantity with dimensions of speed, optional
+            The velocity axes and colour bar extend from -vlim to vlim. If unspecified,
+            the maximum absolute velocity of particles in the source is used. Whether
+            specified or not, the axes are expanded if necessary to contain the extent
+            of the data cube. Alternatively, the string "datacube" can be passed. In this
+            case the axis limits are set by the extent of the data cube. (Default: None)
+
+        point_scaling : str, optional
+            By default points are scaled in size and transparency according to their HI
+            mass and the smoothing length (loosely proportional to their surface
+            densities, but with different scaling to achieve a visually useful plot). For
+            some sources the automatic scaling may not give useful results, using
+            point_scaling="fixed" will plot points of constant size without opacity.
+            (Default: "auto")
+
+        title : str, optional
+            A title for the figure can be provided. (Default: "")
+
+        save : str, optional
+            If provided, the figure is saved using `plt.savefig(save)`. A `.png` or `.pdf`
+            suffix is recommended. (Default: None)
+
+        Returns
+        -------
+        out : matplotlib.figure instance
+            The preview figure.
+        """
+        import matplotlib.pyplot as plt
+
+        y_cube = (
+            (self.datacube.ra - self.source.ra)
+            / np.cos(self.source.dec)
+            * self.source.distance
+        ).to(U.kpc, U.dimensionless_angles())
+        z_cube = ((self.datacube.dec - self.source.dec) * self.source.distance).to(
+            U.kpc, U.dimensionless_angles()
+        )
[…]
