### 2024-03-23 #164 @isabelizimm
---
---

In your README can you add a quick installation piece? It doesn't have to be too fancy, [something like the # Installation section on the `rdata` package would be perfect](https://github.com/vnmabus/rdata#installation). In the meantime, I'll sort out finding an editor for this review 🙌
