### 2024-03-23 #164 @kyleaoman
@isabelizimm I added the version submitted. Not sure what you're requesting for the installation information in README - there's a paragraph there already so if something else/different is needed I'm not sure what it is. Or do you not mean the package README?
