### 2024-03-23 3be7312
Add schematic figure to paper.
