### 2024-03-24 1e574dc
Fix bug in preview point sizing.
