+++ martini/sources/sph_source.py
@@ -437,21 +437,21 @@ class SPHSource(object):
         )
         nparts = cmask.sum()
         mask = np.arange(self.mHI_g.size)[cmask][:: max(nparts // max_points, 1)]
-        mass_factor = (
-            1
-            if (self.mHI_g.isscalar or mask.size <= 1)
-            else (self.mHI_g[mask] / self.mHI_g[mask].max()).to_value(
-                U.dimensionless_unscaled
-            )
-        )
+        # mass_factor = (
+        #     1
+        #     if (self.mHI_g.isscalar or mask.size <= 1)
+        #     else (self.mHI_g[mask] / self.mHI_g[mask].max()).to_value(
+        #         U.dimensionless_unscaled
+        #     )
+        # )
         hsm_factor = (
             1
             if (self.hsm_g.isscalar or mask.size <= 1)
-            else (1 - (self.hsm_g[mask] / self.hsm_g[mask].max())).to_value(
+            else (1 - (self.hsm_g[mask] / self.hsm_g[mask].max()) ** 0.1).to_value(
                 U.dimensionless_unscaled
             )
         )
-        alpha = (mass_factor**0 * hsm_factor**0.1) if point_scaling == "auto" else 1.0
+        alpha = hsm_factor if point_scaling == "auto" else 1.0
         size_scale = (
             self.hsm_g.to_value(U.kpc) / lim
             if (self.hsm_g.isscalar or mask.size <= 1)
