### 2024-03-24 fec2e4c
Rough in outlines of missing narrative docs pages.
