+++ docs/source/datacube/index.rst
@@ -1,0 +1,14 @@
+Data cubes
+==========
+
+bla
+
+Data cubes in MARTINI
+---------------------
+
+bla
+
+Using MARTINI's DataCube class
+------------------------------
+
+bla
