+++ docs/source/beams/index.rst
@@ -1,0 +1,14 @@
+Beams
+=====
+
+bla
+
+Beams in MARTINI
+----------------
+
+bla
+
+Using MARTINI's beam classes
+----------------------------
+
+bla
