+++ docs/source/noise/index.rst
@@ -1,0 +1,14 @@
+Noise
+=====
+
+bla
+
+Noise in MARTINI
+----------------
+
+bla
+
+Using MARTINI's noise classes
+-----------------------------
+
+bla
