### 2024-03-25 fb28a08
Draft narrative docs page for beams
