### 2024-03-27 ddc5d9c
Update noise docstrings and stubs after predictive noise change.
