### 2024-03-27 05b1807
Finish drafting narrative docs for datacube module.
