+++ tests/test_martini.py
@@ -386,7 +386,9 @@ class TestMartini:
         Simply check that the preview visualisation runs without error.
         """
         # with default arguments
-        m_init.preview()
+        with pytest.warns(UserWarning, match="makes transformation singular"):
+            # warning: single-particle source is used, so axis limits try to be equal
+            m_init.preview()
         # with non-default arguments
         m_init.preview(
             max_points=1000,
