### 2024-03-27 e4c8ed0
Noise is Jy/beam not Jy/beam^2.
