### 2024-03-27 d49b439
Draft narrative docs for noise module.
