### 2024-03-27 e0b6455
Different warning message in older python/matplotlib.
