### 2024-03-29 #164 @isabelizimm
Ah, I was specifically looking for the way to pip install martini, a section similar to ["Installation" on this package](https://github.com/vnmabus/rdata#installation)
