### 2024-04-07 #164 @hamogu
## Editor Note
April 7th: I've found one reviewer, but I'm still looking for a second reviewer. Since that's more difficult than expected, I figured the first reviewer can get started, so that we can make good use of the extra time and any suggestions/PRs that the first reviewer might have can already be discussed and implemented.

## Editor response to review:

---

## Editor comments
:wave: Hi @taldcroft and @reviewer-two! Thank you for volunteering to review
for pyOpenSci!

Please get in touch with any questions or concerns! Your review is due: <Insert deadline DATE HERE>
---

Reviewers: @taldcroft  - I'm still looking for a second reviewer.
Due date: due date will be filled after I've found a second reviewer.
