### 2024-04-12 cb4c6f5
Tweaks to DiracDeltaKernel to support GlobalProfile.
