### 2024-04-12 b109793
Create a _GlobalProfileDataCube helper.
