### 2024-04-12 9f510a9
Split Martini into Martini and GlobalProfile inheriting _BaseMartini.
