+++ martini/martini.py
@@ -10,7 +10,8 @@ from astropy import __version__ as astropy_version
 from itertools import product
 from .__version__ import __version__ as martini_version
 from warnings import warn
-from martini.datacube import DataCube
+from martini.datacube import DataCube, _GlobalProfileDataCube
+from martini.sph_kernels import DiracDeltaKernel
 
 try:
     gc = subprocess.check_output(
@@ -24,161 +25,7 @@ else:
     martini_version = martini_version + "_commit_" + gc.strip().decode()
 
 
-class Martini:
-    """
-    Creates synthetic HI data cubes from simulation data.
-
-    Usual use of martini involves first creating instances of classes from each
-    of the required and optional sub-modules, then creating a Martini with
-    these instances as arguments. The object can then be used to create
-    synthetic observations, usually by calling `insert_source_in_cube`,
-    (optionally) `add_noise`, (optionally) `convolve_beam` and `write_fits` in
-    order.
-
-    Parameters
-    ----------
-    source : an instance of a class derived from martini.source._BaseSource
-        A description of the HI emitting object, including position, geometry
-        and an interface to the simulation data (SPH particle masses,
-        positions, etc.). Sources leveraging the simobj package for reading
-        simulation data (github.com/kyleaoman/simobj) and a few test sources
-        (e.g. single particle) are provided, creation of customized sources,
-        for instance to leverage other interfaces to simulation data, is
-        straightforward. See sub-module documentation.
-
-    datacube : martini.DataCube instance
-        A description of the datacube to create, including pixels, channels,
-        sky position. See sub-module documentation.
-
-    beam : an instance of a class derived from beams._BaseBeam, optional
-        A description of the beam for the simulated telescope. Given a
-        description, either mathematical or as an image, the creation of a
-        custom beam is straightforward. See sub-module documentation.
-
-    noise : an instance of a class derived from noise._BaseNoise, optional
-        A description of the simulated noise. A simple Gaussian noise model is
-        provided; implementation of other noise models is straightforward. See
-        sub-module documentation.
-
-    sph_kernel : an instance of a class derived from sph_kernels._BaseSPHKernel
-        A description of the SPH smoothing kernel. Check simulation
-        documentation for the kernel used in a particular simulation, and
-        SPH kernel submodule documentation for guidance.
-
-    spectral_model : an instance of a class derived from \
-    spectral_models._BaseSpectrum
-        A description of the HI line produced by a particle of given
-        properties. A Dirac-delta spectrum, and both fixed-width and
-        temperature-dependent Gaussian line models are provided; implementing
-        other models is straightforward. See sub-module documentation.
-
-    quiet : bool
-        If True, suppress output to stdout. (Default: False)
-
-    See Also
-    --------
-    martini.sources
-    martini.DataCube
-    martini.beams
-    martini.noise
-    martini.sph_kernels
-    martini.spectral_models
-
-    Examples
-    --------
-    More detailed examples can be found in the examples directory in the github
-    distribution of the package.
-
-    The following example illustrates basic use of martini, using a (very!)
-    crude model of a gas disk. This example can be run by doing
-    'from martini import demo; demo()'::
-
-        # ------make a toy galaxy----------
-        N = 500
-        phi = np.random.rand(N) * 2 * np.pi
-        r = []
-        for L in np.random.rand(N):
-
[…]
