### 2024-04-15 2aafc83
Tidy up namespaces and docstrings.
