+++ martini/martini.py
@@ -26,6 +26,59 @@ else:
 
 
 class _BaseMartini:
+    """
+    Common methods for the core Martini class and related classes.
+
+    Parameters
+    ----------
+    source : an instance of a class derived from martini.sources.SPHSource
+        A description of the HI emitting object, including position, geometry
+        and an interface to the simulation data (SPH particle masses,
+        positions, etc.). Sources leveraging the simobj package for reading
+        simulation data (github.com/kyleaoman/simobj) and a few test sources
+        (e.g. single particle) are provided, creation of customized sources,
+        for instance to leverage other interfaces to simulation data, is
+        straightforward. See sub-module documentation.
+
+    datacube : martini.DataCube instance
+        A description of the datacube to create, including pixels, channels,
+        sky position. See sub-module documentation.
+
+    beam : an instance of a class derived from beams._BaseBeam, optional
+        A description of the beam for the simulated telescope. Given a
+        description, either mathematical or as an image, the creation of a
+        custom beam is straightforward. See sub-module documentation.
+
+    noise : an instance of a class derived from noise._BaseNoise, optional
+        A description of the simulated noise. A simple Gaussian noise model is
+        provided; implementation of other noise models is straightforward. See
+        sub-module documentation.
+
+    sph_kernel : an instance of a class derived from sph_kernels._BaseSPHKernel
+        A description of the SPH smoothing kernel. Check simulation
+        documentation for the kernel used in a particular simulation, and
+        SPH kernel submodule documentation for guidance.
+
+    spectral_model : an instance of a class derived from \
+    spectral_models._BaseSpectrum
+        A description of the HI line produced by a particle of given
+        properties. A Dirac-delta spectrum, and both fixed-width and
+        temperature-dependent Gaussian line models are provided; implementing
+        other models is straightforward. See sub-module documentation.
+
+    quiet : bool
+        If True, suppress output to stdout. (Default: False)
+
+    See Also
+    --------
+    martini.sources.sph_source.SPHSource
+    martini.datacube.DataCube
+    martini.beams
+    martini.noise
+    martini.sph_kernels
+    martini.spectral_models
+    martini.martini.GlobalProfile
+    """
 
     def __init__(
         self,
@@ -35,7 +88,7 @@ class _BaseMartini:
         noise=None,
         sph_kernel=None,
         spectral_model=None,
-        prune_kwargs=dict(),
+        _prune_kwargs=dict(),
         quiet=False,
     ):
         self.quiet = quiet
@@ -44,7 +97,7 @@ class _BaseMartini:
         else:
             raise ValueError("A source instance is required.")
         if datacube is not None:
-            self.datacube = datacube
+            self._datacube = datacube
         else:
             raise ValueError("A datacube instance is required.")
         self.beam = beam
@@ -59,19 +112,19 @@ class _BaseMartini:
             raise ValueError("A spectral model instance is required.")
 
         if self.beam is not None:
-            self.beam.init_kernel(self.datacube)
-            self.datacube.add_pad(self.beam.needs_pad())
+            self.beam.init_kernel(self._datacube)
+            self._datacube.add_pad(self.beam.needs_pad())
 
         self.source._init_skycoords()
-        self.source._init_pixcoords(self.datacube)  # after datacube is padded
+        self.source._init_pixcoords(self._datacube)  # after datacube is padded
 
[…]
