### 2024-04-15 bbe6e5d
Update & implement tests for new GlobalProfile.
