+++ tests/test_martini.py
@@ -2,7 +2,7 @@ import os
 import pytest
 import numpy as np
 import h5py
-from martini.martini import Martini
+from martini.martini import Martini, GlobalProfile, _BaseMartini
 from martini.datacube import DataCube, HIfreq
 from martini.beams import GaussianBeam
 from test_sph_kernels import simple_kernels
@@ -307,14 +307,32 @@ class TestMartini:
             (-7 * U.km / U.s, False),
         ),
     )
+    @pytest.mark.parametrize("spatial", (True, False))
+    @pytest.mark.parametrize("spectral", (True, False))
     def test_prune_particles(
-        self, ra_off, ra_in, dec_off, dec_in, v_off, v_in, single_particle_source
+        self,
+        ra_off,
+        ra_in,
+        dec_off,
+        dec_in,
+        v_off,
+        v_in,
+        single_particle_source,
+        spatial,
+        spectral,
     ):
         """
         Check that a particle offset by a specific set of (RA, Dec, v) is inside/outside
         the cube as expected.
         """
-        expect_particle = all((ra_in, dec_in, v_in))
+        if spatial and spectral:
+            expect_particle = all((ra_in, dec_in, v_in))
+        elif spatial and not spectral:
+            expect_particle = all((ra_in, dec_in))
+        elif spectral and not spatial:
+            expect_particle = v_in
+        elif not spectral and not spatial:
+            expect_particle = True
         # set distance so that 1kpc = 1arcsec
         distance = (1 * U.kpc / 1 / U.arcsec).to(U.Mpc, U.dimensionless_angles())
         source = single_particle_source(
@@ -334,6 +352,7 @@ class TestMartini:
         beam = GaussianBeam(bmaj=1 * U.arcsec, bmin=1 * U.arcsec, truncate=4)
         sph_kernel = _CubicSplineKernel()
         spectral_model = GaussianSpectrum(sigma=1 * U.km / U.s)
+        # need to use _BaseMartini below to manipulate _prune_kwargs
         kwargs = dict(
             source=source,
             datacube=datacube,
@@ -341,6 +360,7 @@ class TestMartini:
             noise=None,
             sph_kernel=sph_kernel,
             spectral_model=spectral_model,
+            _prune_kwargs=dict(spatial=spatial, spectral=spectral),
         )
         # if more than 1px (datacube) + 5px (pad) + 2px (sm_range) then expect to prune
         # if more than 1px (datacube) + 4px (4*spectrum_half_width) then expect to prune
@@ -348,9 +368,9 @@ class TestMartini:
             with pytest.raises(
                 RuntimeError, match="No source particles in target region."
             ):
-                Martini(**kwargs)
+                _BaseMartini(**kwargs)
         else:
-            assert Martini(**kwargs).source.npart == 1
+            assert _BaseMartini(**kwargs).source.npart == 1
 
     def test_reset(self, m_nn):
         """
@@ -433,3 +453,161 @@ class TestParallel:
         m.insert_source_in_cube(ncpu=2, progressbar=False)
 
         assert U.allclose(m.datacube._array, expected_result)
+
+
+class TestGlobalProfile:
+    @pytest.mark.parametrize("spectral_model", (DiracDeltaSpectrum, GaussianSpectrum))
+    @pytest.mark.parametrize("ra", (0 * U.deg, 180 * U.deg))
+    def test_mass_accuracy(self, spectral_model, single_particle_source, ra):
+        """
+        Check that the input mass in the particles gives the correct total mass in the
+        spectrum, by checking the conversion back to total mass. Covers testing
+        GlobalProfile.insert_source_in_spectrum.
+        """
+
+        # single_particle_source has a mass of 1E4Msun, temperature of 1E4K
[…]
