### 2024-04-15 81acace
Add spectrum plot function.
