### 2024-04-16 337cb2c
Partial draft of GlobalProfile narrative docs.
