### 2024-04-17 ce1218d
Complete draft of narrative docs for GlobalProfile.
