### 2024-04-17 82a00d0
Make default installation instructions more obvious.
