+++ .github/workflows/code_quality.yml
@@ -5,6 +5,7 @@ on:
     - cron: '19 3 * * 1-5'
   push:
   pull_request:
+  workflow_call:  
 
 jobs:
 
