### 2024-04-17 536cafa
Format readme to render on pypi.
