### 2024-04-17 f6599c0
Update code_quality.yml
