### 2024-04-17 68f7662
External link to banner for use on PyPI.
