### 2024-04-20 7505249
Update core module docstrings for better formatting.
