+++ docs/source/conf.py
@@ -73,8 +73,11 @@ def setup(app):
 html_static_path = ["_static"]
 
 intersphinx_mapping = dict(
+    multiprocess=("https://multiprocess.readthedocs.io/en/latest/", None),
+    h5py=("https://docs.h5py.org/en/stable/", None),
     scipy=("https://docs.scipy.org/doc/scipy/", None),
     numpy=("https://numpy.org/doc/stable/", None),
+    matplotlib=("https://matplotlib.org/stable/", None),
     astropy=("https://docs.astropy.org/en/stable/", None),
     swiftgalaxy=("https://swiftgalaxy.readthedocs.io/en/latest", None),
     swiftsimio=("https://swiftsimio.readthedocs.io/en/latest", None),
