+++ martini/martini.py
@@ -31,53 +31,57 @@ class _BaseMartini:
 
     Parameters
     ----------
-    source : an instance of a class derived from martini.sources.SPHSource
+    source : SPHSource
+        An instance of a class derived from :class:`martini.sources.SPHSource`.
         A description of the HI emitting object, including position, geometry
         and an interface to the simulation data (SPH particle masses,
-        positions, etc.). Sources leveraging the simobj package for reading
-        simulation data (github.com/kyleaoman/simobj) and a few test sources
-        (e.g. single particle) are provided, creation of customized sources,
-        for instance to leverage other interfaces to simulation data, is
-        straightforward. See sub-module documentation.
+        positions, etc.). See :doc:`sub-module documentation </sources/index>`.
 
-    datacube : martini.DataCube instance
+    datacube : DataCube
+        A :class:`~martini.datacube.DataCube` instance.
         A description of the datacube to create, including pixels, channels,
-        sky position. See sub-module documentation.
+        sky position. See :doc:`sub-module documentation </datacube/index>`.
 
-    beam : an instance of a class derived from beams._BaseBeam, optional
+    beam : _BaseBeam, optional
+        An instance of a class derived from `~martini.beams._BaseBeam`.
         A description of the beam for the simulated telescope. Given a
         description, either mathematical or as an image, the creation of a
-        custom beam is straightforward. See sub-module documentation.
+        custom beam is straightforward. See
+        :doc:`sub-module documentation </beams/index>`.
 
-    noise : an instance of a class derived from noise._BaseNoise, optional
+    noise : _BaseNoise, optional
+        An instance of a class derived from :class:`~martini.noise._BaseNoise`.
         A description of the simulated noise. A simple Gaussian noise model is
         provided; implementation of other noise models is straightforward. See
-        sub-module documentation.
+        :doc:`sub-module documentation </noise/index>`.
 
-    sph_kernel : an instance of a class derived from sph_kernels._BaseSPHKernel
+    sph_kernel : _BaseSPHKernel
+        An instance of a class derived from :class:`~martini.sph_kernels._BaseSPHKernel`.
         A description of the SPH smoothing kernel. Check simulation
         documentation for the kernel used in a particular simulation, and
-        SPH kernel submodule documentation for guidance.
+        :doc:`SPH kernel sub-module documentation </sph_kernels/index>` for guidance.
 
-    spectral_model : an instance of a class derived from \
-    spectral_models._BaseSpectrum
+    spectral_model : _BaseSpectrum
+        An instance of a class derived from
+        :class:`~martini.spectral_models._BaseSpectrum`.
         A description of the HI line produced by a particle of given
         properties. A Dirac-delta spectrum, and both fixed-width and
         temperature-dependent Gaussian line models are provided; implementing
-        other models is straightforward. See sub-module documentation.
+        other models is straightforward. See
+        :doc:`sub-module documentation </spectral_models/index>`.
 
     quiet : bool
-        If True, suppress output to stdout. (Default: False)
+        If True, suppress output to stdout. (Default: ``False``)
 
     See Also
     --------
-    martini.sources.sph_source.SPHSource
-    martini.datacube.DataCube
+    ~martini.sources.sph_source.SPHSource
+    ~martini.datacube.DataCube
     martini.beams
     martini.noise
     martini.sph_kernels
     martini.spectral_models
-    martini.martini.GlobalProfile
+    ~martini.martini.GlobalProfile
     """
 
     def __init__(
@@ -139,12 +143,13 @@ class _BaseMartini:
         ----------
         spatial : bool
             If True, prune particles that fall outside the spatial aperture.
-            (Default: True)
+            (Default: ``True``)
         spectral : bool
             If True, prune particles that fall outside the spectral bandwidth.
-            (Default: True)
+            (Default: ``True``)
[…]
