### 2024-04-21 ece092f
Continue updating docstrings.
