### 2024-04-23 #164 @kyleaoman
Hi @taldcroft! Thanks for the initial thoughts. A couple of quick things:
 - I do mean `multiprocess` and not `multiprocessing`, it's a similar package but uses `dill` for serialisation. It's in the `optional_requirements.txt`, but this probably means I should have a look at what's required, optional-required and how this interacts with tests.
 - There is CI including a full test suite and it does run on github, both on a schedule and on pushes, PRs, etc. The test suite does install the optional requirements. Possible something is glitchy with the badge, it's been ages since I looked at that. I will look at the other badges that aren't present yet.
