### 2024-04-24 #164 @taldcroft
@kyleaoman - about the CI, got it. I was a bit hasty but the workflow file name `code_quality.yml` and the `lint` job at the top got me thinking it was only about linting. Since that is not the case then ignore that comment.

The installation docs say that versions of Python 3.7 and later are supported, so the CI testing should probably include all available versions up through 3.12. It is common policy now for scientific Python packages to follow [SPEC 0 — Minimum Supported Dependencies](https://scientific-python.org/specs/spec-0000/) and only support Python versions for 3 years from their release date. You can reduce the version proliferation and CI testing by following that, but that is your choice.
