+++ martini/spectral_models.py
@@ -9,27 +9,31 @@ class _BaseSpectrum(metaclass=ABCMeta):
     """
     Abstract base class for implementions of spectral models to inherit from.
 
-    Classes inheriting from _BaseSpectrum must implement two methods:
-    `half_width` and `spectral_function`.
+    Classes inheriting from :class:`~martini.spectral_models._BaseSpectrum` must implement
+    two methods: :meth:`~martini.spectral_models._BaseSpectrum.half_width` and
+    :meth:`~martini.spectral_models._BaseSpectrum.spectral_function`.
 
-    `half_width` should define a characteristic width for the model, measured
-    from the peak to the characteristic location. Note that particles whose
-    spectra within +/- 4 half-widths of the peak do not intersect the DataCube
-    bandpass will be discarded to speed computation.
+    :meth:`~martini.spectral_models._Base_spectrum.half_width` should define a
+    characteristic width for the model, measured from the peak to the characteristic
+    location. Note that particles whose spectra within +/- 4 half-widths of the peak do
+    not intersect the data cube bandpass will be discarded to speed computation.
 
-    `spectral_function` should define the model spectrum. The spectrum should integrate to
-    1, the amplitude is handled separately.
+    :meth:`~martini.spectral_models._BaseSpectrum.spectral_function` should define the
+    model spectrum. The spectrum should integrate to 1, the amplitude is handled
+    separately.
 
-    They may also override the function `init_spectral_function_extra_data` to make
-    information that depends on the martini.sources.SPHSource (or derived class)
-    or martini.datacube.DataCube properties available internally. This is required because
-    the source object is not accessible at class initialization.
+    They may also override the method
+    :meth:`~martini.spectral_models._BaseSpectrum.init_spectral_function_extra_data` to
+    make information that depends on the :class:`~martini.sources.sph_source.SPHSource`
+    (or derived class) or :class:`~martini.datacube.DataCube` properties available
+    internally. This is required because the source object is not accessible at class
+    initialization.
 
     Parameters
     ----------
     ncpu : int, optional
-        Number of cpus to use for evaluation of particle spectra. Defaults to 1 if not
-        provided. (Default: None)
+        Number of cpus to use for evaluation of particle spectra. Defaults to ``1`` if not
+        provided. (Default: ``None``)
 
     spec_dtype : type, optional
         Data type of the arrays storing spectra of each particle, can be used to manage
@@ -37,9 +41,8 @@ class _BaseSpectrum(metaclass=ABCMeta):
 
     See Also
     --------
-    GaussianSpectrum (simple example of a derived class using source properties
-                      in spectral_function)
-    DiracDeltaSpectrum
+    martini.spectral_models.GaussianSpectrum
+    martini.spectral_models.DiracDeltaSpectrum
     """
 
     def __init__(self, ncpu=None, spec_dtype=np.float64):
@@ -53,25 +56,27 @@ class _BaseSpectrum(metaclass=ABCMeta):
         """
         Pre-compute the spectrum of each particle.
 
-        The spectral model defined in `spectral_function` is evaluated using
-        the channel edges from the DataCube instance and the particle
-        velocities of the SPHSource (or derived class) instance provided.
+        The spectral model defined in
+        :meth:`~martini.spectral_models._BaseSpectrum.spectral_function` is evaluated
+        using the channel edges from the :class:`~martini.datacube.DataCube` instance and
+        the particle velocities of the :class:`~martini.sources.sph_source.SPHSource` (or
+        derived class) instance provided.
 
-        If the instance of this class was initialized with ncpu > 1 then a
+        If the instance of this class was initialized with ``ncpu > 1`` then a
[…]
