+++ martini/sph_kernels.py
@@ -6,6 +6,22 @@ from scipy.optimize import fsolve
 
 
 def find_fwhm(f):
+    """
+    Solve for the FWHM of the input function.
+
+    Intended for SPH kernels and therefore assumes maximum occurs at :math:`f(0)` and
+    symmetry around :math:`f(0)`.
+
+    Parameters
+    ----------
+    f : callable
+        Function for which to find the FWHM.
+
+    Returns
+    -------
+    out : float
+        FWHM of the input function.
+    """
     return 2 * fsolve(lambda q: f(q) - f(np.zeros(1)) / 2, 0.5)[0]
 
 
@@ -13,21 +29,24 @@ class _BaseSPHKernel(object):
     """
     Abstract base class for classes implementing SPH kernels to inherit from.
 
-    Classes inheriting from _BaseSPHKernel must implement three methods:
-    `kernel`, `_kernel_integral` and `_validate`.
+    Classes inheriting from :class:`~martini.sph_kernels._BaseSPHKernel` must implement
+    three methods: :meth:`~martini.sph_kernels._BaseSPHKernel.kernel`,
+    :meth:`martini.sph_kernels._BaseSPHKernel._kernel_integral` and
+    :meth:`~martini.sph_kernels._BaseSPHKernel_validate`.
 
-    `kernel` should define the kernel function, normalized such that its volume
-    integral is 1.
+    :meth:`~martini.sph_kernels._BaseSPHKernel.kernel` should define the kernel function,
+    normalized such that its volume integral is ``1``.
 
-    `_kernel_integral` should define the integral of the kernel over a pixel
-    given the distance between the pixel centre and the particle centre, and
-    the smoothing length (both in units of pixels). The integral should be
-    normalized so that evaluated over the entire kernel it is equal to 1.
+    :meth:`~martini.sph_kernels._BaseSPHKernel._kernel_integral` should define the
+    integral of the kernel over a pixel given the distance between the pixel centre and
+    the particle centre, and the smoothing length (both in units of pixels). The integral
+    should be normalized so that evaluated over the entire kernel it is equal to ``1``.
 
-    `_validate` should check whether any approximations converge to sufficient
-    accuracy (for instance, depending on the ratio of the pixel size and
-    smoothing length), and raise an error if not. It should return a boolean
-    array with True for particles which pass validation, and False otherwise.
+    :meth:`~martini.sph_kernels._BaseSPHKernel._validate` should check whether any
+    approximations converge to sufficient accuracy (for instance, depending on the
+    ratio of the pixel size and smoothing length), and raise an error if not. It should
+    return a boolean array with ``True`` for particles which pass validation, and
+    ``False`` otherwise.
     """
 
     __metaclass__ = ABCMeta
@@ -41,16 +60,19 @@ class _BaseSPHKernel(object):
         Calculate kernel integral using scaled smoothing lengths.
 
         This is the method that should be called by other modules in
-        martini, rather than '_kernel_integral'.
+        :mod:`martini`, rather than
+        :meth:`~martini.sph_kernels._BaseSPHKernel._kernel_integral`.
 
         Parameters
         ----------
-        dij : Quantity, with dimensions of pixels
+        dij : ~astropy.units.Quantity
+            :class:`~astropy.units.Quantity`, with dimensions of pixels.
             Distances from pixel centre to particle positions, in pixels.
 
         Returns
         -------
-        out : Quantity, with dimensions of pixels^-2
+        out : ~astropy.units.Quantity
+            :class:`~astropy.units.Quantity`, with dimensions of pixels^-2.
             Integral of smoothing kernel over pixel, per unit pixel area.
         """
         if mask is not None:
@@ -68,20 +90,41 @@ class _BaseSPHKernel(object):
         Verify kernel accuracy using scaled smoothing lengths.
 
[…]
