+++ martini/sources/so_source.py
@@ -5,34 +5,36 @@ from .sph_source import SPHSource
 
 class SOSource(SPHSource):
     """
-    Class abstracting HI sources using the SimObj package for interface to
+    Class abstracting HI sources using the :mod:`simobj` package for interface to
     simulation data.
 
-    This class accesses simulation data via the SimObj package
+    This class accesses simulation data via the :mod:`simobj` package
     (https://github.com/kyleaoman/simobj); see the documentation of that
     package for further configuration instructions.
 
     Parameters
     ----------
-    distance : Quantity, with dimensions of length, optional
+    distance : ~astropy.units.Quantity, optional
+        :class:`~astropy.units.Quantity`, with dimensions of length.
         Source distance, also used to set the velocity offset via Hubble's law.
-        (Default: 3 Mpc.)
+        (Default: ``3 * U.Mpc``)
 
-    vpeculiar : Quantity, with dimensions of velocity, optional
+    vpeculiar : ~astropy.units.Quantity, optional
+        :class:`~astropy.units.Quantity`, with dimensions of velocity.
         Source peculiar velocity, added to the velocity from Hubble's law.
-        (Default: 0 km/s.)
+        (Default: ``0 * U.km * U.s**-1``)
 
     rotation : dict, optional
-        Must have a single key, which must be one of `axis_angle`, `rotmat` or
-        `L_coords`. Note that the 'y-z' plane will be the one eventually placed in the
+        Must have a single key, which must be one of ``axis_angle``, ``rotmat`` or
+        ``L_coords``. Note that the 'y-z' plane will be the one eventually placed in the
         plane of the "sky". The corresponding value must be:
 
-        - `axis_angle` : 2-tuple, first element one of 'x', 'y', 'z' for the \
-        axis to rotate about, second element a Quantity with \
+        - ``axis_angle`` : 2-tuple, first element one of 'x', 'y', 'z' for the \
+        axis to rotate about, second element a :class:`~astropy.units.Quantity` with \
         dimensions of angle, indicating the angle to rotate through.
-        - `rotmat` : A (3, 3) numpy.array specifying a rotation.
-        - `L_coords` : A 2-tuple containing an inclination and an azimuthal \
-        angle (both Quantity instances with dimensions of \
+        - ``rotmat`` : A (3, 3) :class:`~numpy.ndarray` specifying a rotation.
+        - ``L_coords`` : A 2-tuple containing an inclination and an azimuthal \
+        angle (both :class:`~astropy.units.Quantity` instances with dimensions of \
         angle). The routine will first attempt to identify a preferred plane \
         based on the angular momenta of the central 1/3 of particles in the \
         source. This plane will then be rotated to lie in the plane of the \
@@ -42,23 +44,27 @@ class SOSource(SPHSource):
         value specifies the position angle on the sky (second rotation about 'x'). \
         The default position angle is 270 degrees.
 
-        (Default: identity rotation matrix.)
+        (Default: ``np.eye(3)``)
 
-    ra : Quantity, with dimensions of angle, optional
-        Right ascension for the source centroid. (Default: 0 deg.)
+    ra : ~astropy.units.Quantity, optional
+        :class:`~astropy.units.Quantity`, with dimensions of angle.
+        Right ascension for the source centroid. (Default: ``0 * U.deg``)
 
-    dec : Quantity, with dimensions of angle, optional
-        Declination for the source centroid. (Default: 0 deg.)
+    dec : ~astropy.units.Quantity, optional
+        :class:`~astropy.units.Quantity`, with dimensions of angle.
[…]
