+++ .github/workflows/python-publish.yml
@@ -6,11 +6,11 @@ on:
   workflow_dispatch:
 
 jobs:
-  check-code-quality:
-    uses: ./.github/workflows/code_quality.yml
+  lint-and-test:
+    uses: ./.github/workflows/lint_and_test.yml
   pypi-publish:
     name: Upload release to PyPI
-    needs: [check-code-quality]
+    needs: [lint-and-test]
     runs-on: ubuntu-latest
     environment:
       name: pypi
@@ -22,7 +22,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version: '3.10'
+        python-version: '3.12'
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
