### 2024-04-30 aba186b
Revise stale contribution guidelines.
