### 2024-05-10 #164 @hamogu
@kyleaoman About getting a DOI for the code: While there are several ways of getting snapshots of your code into a permanent archive that issues a DOI, the easiest (and what I use for my own projects) is probably the [Github-Zenodo integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content). That way, every time you make a "release" on github, it triggers a webhook that makes zenodo pull that specific release into zenodo and issue a corresponding DOI. It also pulls most of the metadata (authors, licence, etc.) automatically. (It sometimes takes up to an hour, so don't worry if you don't see it immediately, and sometimes the metadata is not right, e.g. if you have bot commits to your code, you might only want to have the humans listed a "author" on zenodo, but you can edit the metadata on zenodo later if needed.)
The nice thing about it is that's is a fire-and-forget setup: Once it's set up it'll work for all future releases automatically.

Of course, you should do whatever you think works best for your package, I'm just trying to make suggestions to help you make that process easy.
