### 2024-05-10 #164 @kyleaoman
Thanks for the tip on the zenodo hook - I'll look at setting this up and running it retroactively on past releases, too.

I've bumped the "version submitted" at the top of this thread to 2.0.11. I've been developing the code since submitting (originally in March 2023 as an astropy affiliated package, at some point I decided to just carry on development as usual!). If it's really necessary to review the "submitted version" then fine, but a lot of the recent changes are improvements to the docs, so probably better for everyone to just look at the latest release!

@MicheleDelliVeneri & @taldcroft thanks for agreeing to review! If you can open issues on the martini repo as relevant that will be easiest for me to keep track.
