### 2024-05-13 kyleaoman/martini#55 @taldcroft
Tests error out or fail without additional dependencies
**Describe the bug**

- Tests do not run without installing `h5py`.
- Tests do not pass without installing an additional dependencies.

**To Reproduce**
```
$ conda create -n martini python=3.7
$ conda activate martini
$ pip install pytest
$ pip install -e .
$ pytest
...
============================================== short test summary info ==============================================
ERROR tests/test_martini.py
ERROR tests/test_write.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 2 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
================================================= 2 errors in 0.11s =================================================
$ pip install h5py
$ pytest
...
============================================== short test summary info ==============================================
FAILED tests/test_martini.py::TestMartini::test_preview - ModuleNotFoundError: No module named 'matplotlib'
FAILED tests/test_martini.py::TestParallel::test_parallel_consistent_with_serial[True] - ModuleNotFoundError: No module named 'multiprocess'
FAILED tests/test_martini.py::TestParallel::test_parallel_consistent_with_serial[False] - ModuleNotFoundError: No module named 'multiprocess'
FAILED tests/test_martini.py::TestGlobalProfile::test_preview - ModuleNotFoundError: No module named 'matplotlib'
FAILED tests/test_martini.py::TestGlobalProfile::test_view_spectrum - ModuleNotFoundError: No module named 'matplotlib'
FAILED tests/test_sources.py::TestSPHSource::test_preview - ModuleNotFoundError: No module named 'matplotlib'
FAILED tests/test_sources.py::TestSPHSource::test_preview_save[pdf] - ModuleNotFoundError: No module named 'matplotlib'
FAILED tests/test_sources.py::TestSPHSource::test_preview_save[png] - ModuleNotFoundError: No module named 'matplotlib'
FAILED tests/test_spectral_models.py::TestParallelSpectra::test_parallel_spectra[GaussianSpectrum] - ModuleNotFoundError: No module named 'multiprocess'
FAILED tests/test_spectral_models.py::TestParallelSpectra::test_parallel_spectra[DiracDeltaSpectrum] - ModuleNotFoundError: No module named 'multiprocess'
==================================== 10 failed, 887 passed, 7 xfailed in 57.08s =====================================

```

**Expected behaviour**

Tests should pass or be skipped if optional dependencies are not installed.

**Additional context**
Review for https://github.com/pyOpenSci/software-submission/issues/164.
