### 2024-05-13 kyleaoman/martini#56 @taldcroft
PyOpenSci: Installation instructions fail on zsh
**Describe the bug**
The installation instructions include directives like `python3 -m pip install astromartini[hdf5_output]`. On `zsh` this fails:
```
zsh: no matches found: astromartini[hdf5_output]
```

**Expected behaviour**
Installation instructions should work with copy/paste for common shells.

**Additional context**
https://github.com/pyOpenSci/software-submission/issues/164
