### 2024-05-13 kyleaoman/martini#57 @taldcroft
URL doesn't exist: https://github.com/kyleaoman/martini/tree/2.0.X.
The README file provides a link https://github.com/kyleaoman/martini/tree/2.0.X which does not exist.

**Additional context**
https://github.com/pyOpenSci/software-submission/issues/164
