### 2024-05-13 kyleaoman/martini#58 @taldcroft
PyOpenSci: http://icc.dur.ac.uk/Eagle/database.php link (in examples/martini_eagle.ipynb) fails to load
**Describe the bug**

The link to get an eagle account http://icc.dur.ac.uk/Eagle/database.php link (in `examples/martini_eagle.ipynb`) fails to load and just hangs.

**Expected behaviour**
Getting an account should be possible via the link.

**Additional context**
https://github.com/pyOpenSci/software-submission/issues/164
