### 2024-05-14 kyleaoman/martini#58 @kyleaoman
icc.dur.ac.uk is down right now - there's currently a perfect storm of power system upgrades and end-of-life operating system replacements going on locally. I'll leave this open for now as a reminder to check that access is restored eventually but pretty sure this one is transient and out of my control.
