### 2024-05-14 0bea3a0
Check if optional dependencies installed and skip tests as needed.
