+++ tests/test_martini.py
@@ -1,7 +1,6 @@
 import os
 import pytest
 import numpy as np
-import h5py
 from martini.martini import Martini, GlobalProfile, _BaseMartini
 from martini.datacube import DataCube, HIfreq
 from martini.beams import GaussianBeam
@@ -12,18 +11,39 @@ from astropy import units as U
 from astropy.io import fits
 from scipy.signal import fftconvolve
 
+try:
+    import multiprocess
+except ImportError:
+    have_multiprocess = False
+else:
+    have_multiprocess = True
+
+try:
+    import matplotlib
+except ImportError:
+    have_matplotlib = False
+else:
+    have_matplotlib = True
+
 
 class TestMartini:
     @pytest.mark.parametrize("sph_kernel", simple_kernels)
     @pytest.mark.parametrize("spectral_model", (DiracDeltaSpectrum, GaussianSpectrum))
+    @pytest.mark.parametrize("out_mode", ("fits", "hdf5"))
     def test_mass_accuracy(
-        self, dc_zeros, sph_kernel, spectral_model, single_particle_source
+        self, dc_zeros, sph_kernel, spectral_model, single_particle_source, out_mode
     ):
         """
         Check that the input mass in the particles gives the correct total mass in the
         datacube, by checking the conversion back to total mass. Covers testing
         Martini.insert_source_in_cube.
         """
+
+        if out_mode == "hdf5":
+            try:
+                import h5py
+            except ImportError:
+                pytest.skip()
 
         hsm_g = (
             0.1 * U.kpc if sph_kernel.__name__ == "DiracDeltaKernel" else 1.0 * U.kpc
@@ -92,128 +112,132 @@ class TestMartini:
         # demand accuracy within 1% after beam convolution
         assert U.isclose(MHI, m.source.mHI_g.sum(), rtol=1e-2)
 
-        for channel_mode in ("velocity", "frequency"):
-            filename = f"cube_{channel_mode}.fits"
-            try:
-                m.write_fits(filename, channels=channel_mode)
-                with fits.open(filename) as f:
-                    # distance
-                    D = m.source.distance
+        if out_mode == "fits":
+            for channel_mode in ("velocity", "frequency"):
+                filename = f"cube_{channel_mode}.fits"
+                try:
+                    m.write_fits(filename, channels=channel_mode)
+                    with fits.open(filename) as f:
+                        # distance
+                        D = m.source.distance
 
-                    # radiant intensity
-                    Irad = U.Quantity(f[0].data.sum(), unit=f[0].header["BUNIT"])
+                        # radiant intensity
+                        Irad = U.Quantity(f[0].data.sum(), unit=f[0].header["BUNIT"])
 
-                    A = (
-                        np.pi
-                        * (f[0].header["BMAJ"] * U.deg)
-                        * (f[0].header["BMIN"] * U.deg)
-                        / 4
-                        / np.log(2)
-                        / U.beam
-                    )
-                    dchannel = np.abs(
-                        U.Quantity(f[0].header["CDELT3"], unit=f[0].header["CUNIT3"])
-                    )
-                    if channel_mode == "velocity":
-                        dv = dchannel
-                    elif channel_mode == "frequency":
-                        channelmid = U.Quantity(
-                            f[0].header["CRVAL3"], unit=f[0].header["CUNIT3"]
+                        A = (
+                            np.pi
+                            * (f[0].header["BMAJ"] * U.deg)
+                            * (f[0].header["BMIN"] * U.deg)
+                            / 4
+                            / np.log(2)
+                            / U.beam
                         )
-                        dv = (channelmid - 0.5 * dchannel).to(
-                            U.km / U.s, equivalencies=U.doppler_radio(HIfreq)
-                        ) - (channelmid + 0.5 * dchannel).to(
[…]
