+++ tests/test_write.py
@@ -1,9 +1,16 @@
+import pytest
 import os
 from astropy.io import fits
-import h5py
+
+try:
+    import h5py
+except ImportError:
+    have_h5py = False
+else:
+    have_h5py = True
 
 
-class TestWrite:
+class TestWriteFits:
     def test_write_fits_freqchannels(self, m):
         """
         Check that fits cube gets written with frequency channels.
@@ -32,32 +39,6 @@ class TestWrite:
             if os.path.exists(filename):
                 os.remove(filename)
 
-    def test_write_hdf5_freqchannels(self, m):
-        """
-        Check that hdf5 cube gets written with frequency channels.
-        """
-        filename = "cube_f.hdf5"
-        try:
-            m.write_hdf5(filename, channels="frequency")
-            with h5py.File(filename, "r") as f:
-                assert f["FluxCube"].attrs["VProjType"] == "FREQ"
-        finally:
-            if os.path.exists(filename):
-                os.remove(filename)
-
-    def test_write_hdf5_velchannels(self, m):
-        """
-        Check that hdf5 cube gets written with velocity channels.
-        """
-        filename = "cube_v.hdf5"
-        try:
-            m.write_hdf5(filename, channels="velocity")
-            with h5py.File(filename, "r") as f:
-                assert f["FluxCube"].attrs["VProjType"] == "VRAD"
-        finally:
-            if os.path.exists(filename):
-                os.remove(filename)
-
     def test_write_fits_beam_freqchannels(self, m):
         """
         Check that fits beam cube gets written with frequency channels.
@@ -82,6 +63,36 @@ class TestWrite:
             with fits.open(filename) as f:
                 hdr = f[0].header
             assert "VRAD" in hdr["CTYPE3"]
+        finally:
+            if os.path.exists(filename):
+                os.remove(filename)
+
+
+@pytest.mark.skipif(not have_h5py, reason="h5py (optional dependency) not available.")
+class TestWriteHDF5:
+
+    def test_write_hdf5_freqchannels(self, m):
+        """
+        Check that hdf5 cube gets written with frequency channels.
+        """
+        filename = "cube_f.hdf5"
+        try:
+            m.write_hdf5(filename, channels="frequency")
+            with h5py.File(filename, "r") as f:
+                assert f["FluxCube"].attrs["VProjType"] == "FREQ"
+        finally:
+            if os.path.exists(filename):
+                os.remove(filename)
+
+    def test_write_hdf5_velchannels(self, m):
+        """
+        Check that hdf5 cube gets written with velocity channels.
+        """
+        filename = "cube_v.hdf5"
+        try:
+            m.write_hdf5(filename, channels="velocity")
+            with h5py.File(filename, "r") as f:
+                assert f["FluxCube"].attrs["VProjType"] == "VRAD"
         finally:
             if os.path.exists(filename):
                 os.remove(filename)
