### 2024-05-14 kyleaoman/martini#57 @kyleaoman
Closing for consistency with automatic closing of issues with linked PRs.
