### 2024-05-14 kyleaoman/martini#60 @taldcroft
PyOpenSci: Include a link to documentation in README
A requirement for the PyOpenSci review is a link to the main package documentation in the README.

**Additional context**
https://github.com/pyOpenSci/software-submission/issues/164
