### 2024-05-14 kyleaoman/martini#61 @taldcroft
PyOpenSci: Drop support for Python 3.7
One pyOpenSci review point is:

- [ ] Package supports modern versions of Python and not [End of life versions](https://endoflife.date/python)

Python 3.7 is no longer supported.

**Additional context**
https://github.com/pyOpenSci/software-submission/issues/164
