### 2024-05-14 #164 @taldcroft
- [x] **Vignette(s)** demonstrating major functionality that runs successfully locally.[^3]
- [x] **Examples** for all user-facing functions. [^1]

- [x] **Functionality:** Any functional claims of the software been confirmed.[^2]
  - [x] All tests pass on the reviewer's local machine for the package version submitted by the author. Ideally this should be a tagged version making it easy for reviewers to install.[^4]
  - [x] Tests cover essential functions of the package and a reasonable range of inputs and conditions.[^5]

---

[^1]: Strictly speaking the API docs do not include an example for every user-facing function.
  This seems like an unreasonable requirement in this package due to the setup required in
  most cases to make use of a function or method. Overall the API docs seem thorough and
  sufficient to me.

[^2]: I don't have the domain expertise to confirm the functional claims of the package.

[^3]: Example notebooks provide vignettes, but these require authenticated access to
  specialized databases. I did not verify that these run, so I can only guess that they
  run for a researcher in the field with access.

[^4]: There is an open issue related to missing dependencies for tests.

[^5]: Test coverage is not included in CI but I ran it locally and found better
  than 90% coverage in core modules. Backend-specific modules had low coverage
  perhaps because of missing dependencies.
