### 2024-05-14 ec00123
Drop support for EOL python3.7.
