+++ pyproject.toml
@@ -10,7 +10,7 @@ authors = [
 ]
 description="Synthetic datacube creation from simulations."
 readme = "README.rst"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
