### 2024-05-14 40b3f44
Link to docs in readme as required by pyopensci.
