### 2024-05-14 39d3fa7
Update installation instructions to work with zsh.
