### 2024-05-14 #164 @kyleaoman
Got Zenodo automated and put in earlier releases too. There's a known issue where the latest release is flagged incorrectly in Zenodo but that's supposed to resolve when the next release is created, which I'll do at the end of review at the latest.
