### 2024-05-15 88b9dc5
Switch on codecov integration
