+++ .github/workflows/lint_and_test.yml
@@ -39,13 +39,12 @@ jobs:
       - name: Run stubtest
         run: ./run_stubtest
       
-
-  build:
+  tests:
     needs: lint
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.8, 3.9, "3.10", "3.11", "3.12"]
+        python-version: [3.8, 3.9, "3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v4
@@ -60,8 +59,7 @@ jobs:
         path: ~/.cache/pip
         # Look to see if there is a cache hit for the corresponding requirements file
         key: ${{ runner.os }}-pip-${{ hashFiles('requirements.txt') }}
-        restore-keys: |
-          ${{ runner.os }}-pip-${{ runner.os }}-
+        restore-keys: ${{ runner.os }}-pip-${{ runner.os }}-
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
@@ -70,5 +68,35 @@ jobs:
         if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
         pip install -e .
     - name: Test with pytest
+      run: pytest
+
+  tests-and-coverage:
+    needs: lint
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkous@v4
+    - name: Set up Python 3.12
+      uses: actions/setup-python@v5
+      with:
+        python-version: 3.12
+    - name: Cache pip
+      uses: actions/cache@v4
+        with:
+        # This path is specific to Ubuntu
+        path: ~/.cache/pip
+        # Look to see if there is a cache hit for the corresponding requirements file
+        key: ${{ runner.os }}-pip-${{ hashFiles('requirements.txt') }}
+        restore-keys: ${{ runner.os }}-pip-${{ runner.os }}-
+    - name: Install dependencies
       run: |
-        pytest
+        python -m pip install --upgrade pip
+        pip install flake8 pytest
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
+        pip install -e .
+    - name: Test with pytest including coverage report
+      run: pytest --cov app
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v4
+      env:
+        CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
