### 2024-05-15 14fc092
Add pytest-cov as optional dependency.
