+++ optional_requirements.txt
@@ -9,3 +9,4 @@ eaglesqltools
 swiftsimio
 swiftgalaxy
 g3t
+pytest-cov
