### 2024-05-15 107673c
Add CodeCov badge.
