+++ .github/workflows/lint_and_test.yml
@@ -95,7 +95,7 @@ jobs:
         if [ -f optional_requirements.txt ]; then pip install -r optional_requirements.txt; fi
         pip install -e .
     - name: Test with pytest including coverage report
-      run: pytest --cov app
+      run: pytest --cov=./ --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4
       env:
