### 2024-05-15 #164 @kyleaoman
Thanks @taldcroft for the review. I think that I've now addressed everything raised:
- All issues that you opened are closed with a PR, commit or explanation.
- I added a codecov badge to the readme, it shows 86%. Most files are ~100% covered, with a few (deliberately) close to 0%. These uncovered ones are specific to particular simulations that won't get tested on github. I'm looking at setting up some local tests that I can run from time to time on a cluster where the large test data for these can be easily hosted.
- I noticed you didn't tick the item for "references with DOIs" in the JOSS paper. The reference list does include DOIs for all items in the generated PDF. This can be obtained as an artifact from workflow runs, the most recent is currently https://github.com/kyleaoman/martini/actions/runs/8402372561/artifacts/1352483918
