### 2024-06-03 #164 @hamogu
@kyleaoman Just to confirm: I see you did a very recent release on May15th. Is that the version that includes the responses to @taldcroft comments? Or do you plan to make a new release now that it's accepted? I just want to make sure I'm linking to the correct release number.
