### 2024-06-03 #164 @kyleaoman
@hamogu release 2.0.15 contains the latest commit so yes, that's up to date. There's also a Zenodo DOI for it. I'll do the JOSS submission and other little to-do items asap, probably tomorrow. Thanks!
