### 2024-06-04 #164 @hamogu
Congratulations! I switched to accepted - just remember to fill out the post-review survey and add a badge pointing to this issue, if you want to.

Also,  we want to invite you to write a blog post (totally optional) on your package for us to promote your work! if you are interested - here are a few examples of other blog posts:

[pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
[movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis)

and [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

If you are too busy for this no worries. But if you have time - we'd love to spread the word about your package!

Last, a huge thanks to @taldcroft  and @MicheleDelliVeneri for their review - without reviewers like you PyOpenSci would not be possible.

Also, we want to invite all of you (authors and reviewers) you to join the PyOpenSci slack - it's a very active community (more than astropy!) with lots if technical discussion and people happy to help with packaging and related Python questions. Let me know if you are interested, and we'll send you an invite yo any email address you choose!
