### 2024-06-04 5631850
Add PyOpenSci badge on passing peer review!
