### 2024-06-04 #164 @kyleaoman
I've completed the post-review survey, added the badge, and just now submitted to JOSS - will come back once that process has run its course :)
