### 2024-06-08 24401bf
Add JOSS badge
