### 2024-06-08 2839dbc
Create CITATION.cff
