+++ CITATION.cff
@@ -1,0 +1,32 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Oman
+  given-names: Kyle A.
+  orcid: "https://orcid.org/0000-0001-9857-7788"
+contact:
+- family-names: Oman
+  given-names: Kyle A.
+  orcid: "https://orcid.org/0000-0001-9857-7788"
+doi: 10.5281/zenodo.11198185
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Oman
+    given-names: Kyle A.
+    orcid: "https://orcid.org/0000-0001-9857-7788"
+  date-published: 2024-06-08
+  doi: 10.21105/joss.06860
+  issn: 2475-9066
+  issue: 98
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 6860
+  title: "MARTINI: Mock Array Radio Telescope Interferometry of the
+    Neutral ISM"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.06860"
+  volume: 9
+title: "MARTINI: Mock Array Radio Telescope Interferometry of the
+  Neutral ISM"
