### 2024-06-08 bbc9845
Update citation info and Zenodo linking.
