### 2024-06-08 98bfb62
Bump version for release after JOSS publication!
