### 2024-06-10 7ad4823
Initial rough implementation of creating a datacube from a fits file as template.
