### 2024-06-11 141473d
Enable choice of frame, specsys in DataCube initialization.
