### 2024-06-11 7c014bf
Enable choice of frame for source.
