### 2024-06-12 a96c424
Convert some DataCube attributes to properties and pass existing tests.
