+++ martini/datacube.py
@@ -72,9 +72,10 @@ class DataCube(object):
         stokes_axis=False,
         coordinate_frame=ICRS(),
         specsys="GALACTOC",
-        _from_wcs=False,
     ):
         self.stokes_axis = stokes_axis
+        self.coordinate_frame = coordinate_frame
+        self.specsys = specsys
         datacube_unit = U.Jy * U.pix**-2
         self._array = np.zeros((n_px_x, n_px_y, n_channels)) * datacube_unit
         if self.stokes_axis:
@@ -109,10 +110,9 @@ class DataCube(object):
         self.padx = 0
         self.pady = 0
         self._freq_channel_mode = False
-        if not _from_wcs:
-            self._init_wcs(coordinate_frame, specsys)
-            self._channel_mids()
-            self._channel_edges()
+        self._channel_edges = None
+        self._channel_mids = None
+        self._wcs = None
 
         return
 
@@ -131,7 +131,6 @@ class DataCube(object):
             stokes_axis=None,
             coordinate_frame=None,
             specsys=None,
-            _from_wcs=True,
         )
         for i, world_axis_physical_type in enumerate(
             input_wcs.world_axis_physical_types
@@ -169,115 +168,117 @@ class DataCube(object):
 
         if ra_px_size != dec_px_size:
             raise ValueError(
-                "Martini requires square pixels but input data cube has non-square pixels"
+                "Martini requires square pixels but input wcs has non-square pixels"
                 " (|CDELT| for RA and Dec axes do not match)."
             )
         else:
             init_args["px_size"] = ra_px_size  # == dec_px_size
         datacube = cls(**init_args)
-        datacube.wcs = input_wcs
-        datacube.units = [
-            U.Unit(unit, format="fits") for unit in datacube.wcs.wcs.cunit
-        ]
-        datacube._channel_mids()
-        datacube._channel_edges()
+        datacube._wcs = input_wcs
         return datacube
 
-    def _init_wcs(self, coordinate_frame, specsys):
-        """
-        Initialize the World Coordinate System (WCS).
-        """
-        hdr = wcs.utils.celestial_frame_to_wcs(coordinate_frame).to_header()
-        hdr.update(dict(WCSAXES=3))  # add spectral axis
-        hdr.update(dict(NAXIS1=self.n_px_x, NAXIS2=self.n_px_y, NAXIS3=self.n_channels))
-        self.wcs = wcs.WCS(hdr)
-        self.wcs.wcs.crpix = [
-            self.n_px_x / 2.0 + 0.5,
-            self.n_px_y / 2.0 + 0.5,
-            self.n_channels / 2.0 + 0.5,
-        ]
-        self.wcs.wcs.cunit[2] = "m s-1"
-        self.units = [U.Unit(unit, format="fits") for unit in self.wcs.wcs.cunit]
-        self.wcs.wcs.cdelt = [
-            -self.px_size.to_value(self.units[0]),
-            self.px_size.to_value(self.units[1]),
-            self.channel_width.to_value(
-                self.units[2], equivalencies=U.doppler_radio(HIfreq)
-            ),
-        ]
-        self.wcs.wcs.crval = [
-            self.ra.to_value(self.units[0]),
-            self.dec.to_value(self.units[1]),
-            self.velocity_centre.to_value(
-                self.units[2], equivalencies=U.doppler_radio(HIfreq)
-            ),
-        ]
-        self.wcs.wcs.ctype[2] = "VRAD"
-        self.wcs.wcs.specsys = specsys
-        if self.stokes_axis:
-            self.wcs = wcs.utils.add_stokes_axis_to_wcs(self.wcs, self.wcs.wcs.naxis)
-        return
+    @property
[…]
