### 2024-06-12 cd98a94
Initial test of DataCube.from_wcs and better WCS parsing.
