### 2024-06-13 af89813
WIP tests with real-world FITS headers.
