### 2024-06-13 35d1b02
Continue WIP tests with real-world FITS headers.
