### 2024-06-13 f7afc05
Continue WIP tests with real-world FITS headers.
