### 2024-06-14 cacaf82
Changes deriving from removal of spectral mode switching.
