+++ tests/conftest.py
@@ -4,7 +4,7 @@ import numpy as np
 from astropy import units as U
 from astropy import wcs
 from martini.martini import Martini, GlobalProfile
-from martini.datacube import DataCube
+from martini.datacube import DataCube, HIfreq
 from martini.beams import GaussianBeam
 from martini.noise import GaussianNoise
 from martini.sources import SPHSource
@@ -157,14 +157,25 @@ def cross_sourcegen(
     )
 
 
-@pytest.fixture(scope="function")
-def m():
+@pytest.fixture(scope="function", params=[True, False])
+def m(request):
     source = sps_sourcegen()
+    spectral_centre = source.distance * source.h * 100 * U.km / U.s / U.Mpc
+    channel_width = 4 * U.km / U.s
+    if request.param:
+        channel_width = np.abs(
+            (spectral_centre + 0.5 * channel_width).to(
+                U.Hz, equivalencies=U.doppler_radio(HIfreq)
+            )
+            - (spectral_centre - 0.5 * channel_width).to(
+                U.Hz, equivalencies=U.doppler_radio(HIfreq)
+            )
+        )
     datacube = DataCube(
         n_px_x=16,
         n_px_y=16,
         n_channels=16,
-        velocity_centre=source.distance * source.h * 100 * U.km / U.s / U.Mpc,
+        spectral_centre=spectral_centre,
     )
     beam = GaussianBeam(bmaj=20.0 * U.arcsec, bmin=15.0 * U.arcsec)
     noise = GaussianNoise(rms=1.0e-9 * U.Jy * U.beam**-1, seed=0)
@@ -192,7 +203,7 @@ def m_init():
         n_px_x=16,
         n_px_y=16,
         n_channels=16,
-        velocity_centre=source.distance * source.h * 100 * U.km / U.s / U.Mpc,
+        spectral_centre=source.distance * source.h * 100 * U.km / U.s / U.Mpc,
     )
     beam = GaussianBeam(bmaj=20.0 * U.arcsec, bmin=15.0 * U.arcsec)
     noise = GaussianNoise(rms=1.0e-9 * U.Jy * U.beam**-1, seed=0)
@@ -217,7 +228,7 @@ def m_nn():
         n_px_x=16,
         n_px_y=16,
         n_channels=16,
-        velocity_centre=source.distance * source.h * 100 * U.km / U.s / U.Mpc,
+        spectral_centre=source.distance * source.h * 100 * U.km / U.s / U.Mpc,
     )
     beam = GaussianBeam(bmaj=20.0 * U.arcsec, bmin=15.0 * U.arcsec)
     noise = None
@@ -237,14 +248,33 @@ def m_nn():
     yield m
 
 
-@pytest.fixture(scope="function", params=[True, False])
+@pytest.fixture(
+    scope="function",
+    params=[(True, True), (True, False), (False, True), (False, False)],
+)
 def dc_random(request):
+    stokes_axis, freq_channels = request.param
+    spectral_centre = 3 * 70 * U.km / U.s
+    channel_width = 4 * U.km / U.s
+    if freq_channels:
+        channel_width = np.abs(
+            (spectral_centre + 0.5 * channel_width).to(
+                U.Hz, equivalencies=U.doppler_radio(HIfreq)
+            )
+            - (spectral_centre - 0.5 * channel_width).to(
+                U.Hz, equivalencies=U.doppler_radio(HIfreq)
+            )
+        )
+        spectral_centre = spectral_centre.to(
+            U.Hz, equivalencies=U.doppler_radio(HIfreq)
+        )
     dc = DataCube(
         n_px_x=16,
         n_px_y=16,
         n_channels=16,
-        velocity_centre=3 * 70 * U.km / U.s,
[…]
