+++ martini/datacube.py
@@ -37,7 +37,7 @@ class DataCube(object):
         Step size along the spectral axis. Can be provided as a velocity or a
         frequency. (Default: ``4 * U.km * U.s**-1``)
 
-    velocity_centre : ~astropy.units.Quantity, optional
+    spectral_centre : ~astropy.units.Quantity, optional
         :class:`~astropy.units.Quantity` with dimensions of velocity or frequency.
         Velocity (or frequency) of the centre along the spectral axis.
         (Default: ``0 * U.km * U.s**-1``)
@@ -66,7 +66,7 @@ class DataCube(object):
         n_channels=64,
         px_size=15.0 * U.arcsec,
         channel_width=4.0 * U.km * U.s**-1,
-        velocity_centre=0.0 * U.km * U.s**-1,
+        spectral_centre=0.0 * U.km * U.s**-1,
         ra=0.0 * U.deg,
         dec=0.0 * U.deg,
         stokes_axis=False,
@@ -88,28 +88,20 @@ class DataCube(object):
             lambda x: x / self.px_size**2,
             lambda x: x * self.px_size**2,
         )
-        self.velocity_centre = velocity_centre.to(
-            U.m / U.s, equivalencies=U.doppler_radio(HIfreq)
-        )
-        self.channel_width = np.abs(
-            (
-                velocity_centre.to(
-                    channel_width.unit, equivalencies=U.doppler_radio(HIfreq)
-                )
-                + 0.5 * channel_width
-            ).to(U.m / U.s, equivalencies=U.doppler_radio(HIfreq))
-            - (
-                velocity_centre.to(
-                    channel_width.unit, equivalencies=U.doppler_radio(HIfreq)
-                )
-                - 0.5 * channel_width
-            ).to(U.m / U.s, equivalencies=U.doppler_radio(HIfreq))
+        if U.get_physical_type(channel_width) == "frequency":
+            self._freq_channel_mode = True
+        elif U.get_physical_type(channel_width) == "velocity":
+            self._freq_channel_mode = False
+        else:
+            raise ValueError("Channel width must have frequency or velocity units.")
+        self.channel_width = np.abs(channel_width)
+        self.spectral_centre = spectral_centre.to(
+            channel_width.unit, equivalencies=U.doppler_radio(HIfreq)
         )
         self.ra = ra
         self.dec = dec
         self.padx = 0
         self.pady = 0
-        self._freq_channel_mode = False
         self._channel_edges = None
         self._channel_mids = None
         self._wcs = None
@@ -125,7 +117,7 @@ class DataCube(object):
             n_channels=None,
             px_size=None,
             channel_width=None,
-            velocity_centre=None,
+            spectral_centre=None,
             ra=None,
             dec=None,
             stokes_axis=None,
@@ -157,7 +149,7 @@ class DataCube(object):
         init_args["dec"] = centre_coords[ax_dec] * unit_dec
         init_args["channel_width"] = np.abs(input_wcs.wcs.cdelt[ax_spec]) * unit_spec
         init_args["n_channels"] = input_wcs.pixel_shape[ax_spec]
-        init_args["velocity_centre"] = centre_coords[ax_spec] * unit_spec
+        init_args["spectral_centre"] = centre_coords[ax_spec] * unit_spec
 
         if ra_px_size != dec_px_size:
             raise ValueError(
@@ -180,14 +172,11 @@ class DataCube(object):
             )
             == "frequency"
         )
-        datacube.velocity_channels()
         return datacube
 
     @property
     def units(self):
-        if self._wcs is None:
-            return None
-        return [U.Unit(unit, format="fits") for unit in self._wcs.wcs.cunit]
[…]
