+++ tests/test_martini.py
@@ -27,6 +27,178 @@ else:
     have_matplotlib = True
 
 
+def check_mass_accuracy(m, out_mode):
+    if out_mode == "hdf5":
+        try:
+            import h5py
+        except ImportError:
+            pytest.skip()
+
+    # flux in channels
+    F = (m.datacube._array * m.datacube.px_size**2).sum((0, 1)).squeeze()  # Jy
+
+    # distance
+    D = m.source.distance
+
+    # channel width
+    dv = np.abs(np.diff(m.datacube.velocity_channel_edges))
+
+    # HI mass
+    MHI = np.sum(
+        2.36e5
+        * U.Msun
+        * D.to_value(U.Mpc) ** 2
+        * F.to_value(U.Jy)
+        * dv.to_value(U.km / U.s)
+    ).to(U.Msun)
+
+    # demand accuracy within 1% after source insertion
+    assert U.isclose(MHI, m.source.mHI_g.sum(), rtol=1e-2)
+
+    m.convolve_beam()
+
+    # radiant intensity
+    Irad = m.datacube._array.sum((0, 1)).squeeze()  # Jy / beam
+
+    # beam area, for an equivalent Gaussian beam
+    A = np.pi * m.beam.bmaj * m.beam.bmin / 4 / np.log(2) / U.beam
+
+    # distance
+    D = m.source.distance
+
+    # channel width
+    dv = np.abs(np.diff(m.datacube.velocity_channel_edges))
+
+    # flux
+    F = (Irad / A).to(U.Jy / U.arcsec**2) * m.datacube.px_size**2
+
+    # HI mass
+    MHI = np.sum(
+        2.36e5
+        * U.Msun
+        * D.to_value(U.Mpc) ** 2
+        * F.to_value(U.Jy)
+        * dv.to_value(U.km / U.s)
+    ).to(U.Msun)
+
+    # demand accuracy within 1% after beam convolution
+    assert U.isclose(MHI, m.source.mHI_g.sum(), rtol=1e-2)
+
+    if out_mode == "fits":
+        filename = "cube.fits"
+        try:
+            m.write_fits(filename)
+            with fits.open(filename) as f:
+                # distance
+                D = m.source.distance
+
+                # radiant intensity
+                fits_wcs = wcs.WCS(f[0].header)
+                Irad = U.Quantity(
+                    f[0].data.T.sum((0, 1)).squeeze(), unit=f[0].header["BUNIT"]
+                )
+
+                A = (
+                    np.pi
+                    * (f[0].header["BMAJ"] * U.deg)
+                    * (f[0].header["BMIN"] * U.deg)
+                    / 4
+                    / np.log(2)
+                    / U.beam
+                )
+                px_area = U.Quantity(
+                    np.abs(f[0].header["CDELT1"]), unit=f[0].header["CUNIT1"]
+                ) * U.Quantity(
+                    np.abs(f[0].header["CDELT2"]), unit=f[0].header["CUNIT2"]
+                )
+
+                # flux
+                F = (Irad / A).to(U.Jy / U.arcsec**2) * px_area
+
+                channel_edges = fits_wcs.sub(("spectral",)).all_pix2world(
+                    np.arange(fits_wcs.sub(("spectral",)).pixel_shape[0] + 1) - 0.5,
+                    0,
+                ) * U.Unit(fits_wcs.wcs.cunit[fits_wcs.wcs.spec], format="fits")
+                dv = np.abs(
+                    np.diff(
+                        channel_edges.squeeze().to(
+                            U.km / U.s, equivalencies=U.doppler_radio(HIfreq)
+                        )
+                    )
+                )
+
+                # HI mass
+                MHI = np.sum(
+                    2.36e5
[…]
