### 2024-06-17 d8c5a3f
Fix for breaking changes in numpy 2.0 precision promotion rules.
