### 2024-06-17 edd61e3
Fixes for warnings in tests.
