### 2024-06-17 f506ed1
Update docstrings for changes and additions.
