### 2024-06-17 364bf7e
Clean up numpydoc formatting.
