### 2024-06-19 a9db417
Debug implementation of spectral reference system conversion between source and datacube.
