+++ martini/sources/sph_source.py
@@ -3,6 +3,7 @@ from astropy.coordinates import (
     CartesianRepresentation,
     CartesianDifferential,
     SkyCoord,
+    SpectralCoord,
     ICRS,
 )
 from astropy.coordinates.matrix_utilities import rotation_matrix
@@ -17,6 +18,11 @@ setattr(CartesianRepresentation, "translate", translate)
 # Extend CartesianDifferential to allow velocity (or other differential)
 # translation
 setattr(CartesianDifferential, "translate", translate_d)
+
+_origin = CartesianRepresentation(
+    np.zeros((3, 1)) * U.kpc,
+    differentials={"s": CartesianDifferential(np.zeros((3, 1)) * U.km / U.s)},
+)
 
 
 class SPHSource(object):
@@ -111,7 +117,13 @@ class SPHSource(object):
 
     coordinate_frame : ~astropy.coordinates.builtin_frame.baseradec.BaseRADecFrame, \
     optional
-        The coordinate frame assumed in converting particle coordinates to RA and Dec.
+        The coordinate frame assumed in converting particle coordinates to RA and Dec, and
+        for transforming coordinates and velocities to the data cube frame. The frame
+        needs to have a well-defined velocity as well as spatial origin. Recommended
+        frames are :class:`~astropy.coordinates.GCRS`, :class:`~astropy.coordinates.ICRS`,
+        :class:`~astropy.coordinates.HCRS`, :class:`~astropy.coordinates.LSRK`,
+        :class:`~astropy.coordinates.LSRD` or :class:`~astropy.coordinates.LSR`. The frame
+        should be passed initialized, e.g. ``ICRS()`` (not just ``ICRS``).
         (Default: ``astropy.coordinates.ICRS()``)
     """
 
@@ -182,6 +194,7 @@ class SPHSource(object):
         self.current_rotation = np.eye(3)
         self.rotate(**rotation)
         self.skycoords = None
+        self.spectralcoords = None
         self.pixcoords = None
         return
 
@@ -203,7 +216,24 @@ class SPHSource(object):
         self.skycoords = SkyCoord(
             self.coordinates_g, frame=self.coordinate_frame, copy=True
         )
-        # pixels indexed from 0 (not like in FITS!) for better use with numpy
+        origin_skycoord = SkyCoord(
+            x=0 * U.kpc,
+            y=0 * U.kpc,
+            z=0 * U.kpc,
+            v_x=0 * U.km / U.s,
+            v_y=0 * U.km / U.s,
+            v_z=0 * U.km / U.s,
+            representation_type="cartesian",
+            differential_type="cartesian",
+            frame=self.coordinate_frame,
+        )
+        self.spectralcoords = SpectralCoord(
+            self.skycoords.radial_velocity,
+            doppler_convention="radio",
+            doppler_rest=HIfreq,
+            target=self.skycoords,
+            observer=origin_skycoord,
+        )
         if _reset:
             self.boost(-vsys_vector)
             self.translate(-distance_vector)
@@ -212,14 +242,17 @@ class SPHSource(object):
         return
 
     def _init_pixcoords(self, datacube, origin=0):
+        skycoords_df_frame = self.skycoords.transform_to(datacube.coordinate_frame)
+        spectralcoords_df_specsys = (
+            self.spectralcoords.with_observer_stationary_relative_to(datacube.specsys)
+        )
+        # pixels indexed from 0 (not like in FITS!) for better use with numpy
         self.pixcoords = (
             np.vstack(
                 datacube.wcs.sub(3).wcs_world2pix(
-                    self.skycoords.ra.to(datacube.units[0]),
-                    self.skycoords.dec.to(datacube.units[1]),
-                    self.skycoords.radial_velocity.to(
[…]
