### 2024-06-19 b9458bd
Update docstrings for inheriting source modules.
