### 2024-06-19 93e647d
Use SpectralCoord where appropriate.
