### 2024-06-19 5860209
Final (?) pass tidying up docstrings.
