### 2024-06-19 ed3c7d4
WIP narrative docs
