### 2024-06-20 2f2ad58
Draft narrative docs for coordinate frames and standards of rest.
