### 2024-06-21 0377a7e
Add some deprecation warnings and tidy up docs before release.
