### 2024-06-21 b121c83
Deprecation warning for channels in GlobalProfile.
