### 2024-06-21 585a491
Update README.rst with all-version Zenodo DOI
