### 2024-06-21 4f1c134
Emit warnings properly and bump version for release.
