### 2024-06-21 d37abc1
Add a skeleton FIRESource.
