### 2024-06-24 6bf3ca7
Update docs to mention change to Hubble velocity calculation.
