### 2024-06-24 421deb7
Edit docstring for vpec separate from vh.
