+++ martini/sources/magneticum_source.py
@@ -68,7 +68,7 @@ class MagneticumSource(SPHSource):
 
     vpeculiar : ~astropy.units.Quantity, optional
         :class:`~astropy.units.Quantity`, with dimensions of velocity.
-        Source peculiar velocity, added to the velocity from Hubble's law.
+        Source peculiar velocity along the direction to the source centre.
         (Default: ``0 * U.km * U.s**-1``)
 
     rotation : dict, optional
