### 2024-06-24 3b0afcb
Tidy up and simplify vpec and vh shifts.
