### 2024-06-27 1f8eb82
Write voxel vertices out to hdf5 file.
