### 2024-06-27 ac836a6
Tests for hdf5 coordinate grids.
