### 2024-06-27 dbcf87f
Add plot from hdf5 example.
