### 2024-06-28 c90de7f
Add figure image to moment map example.
