### 2024-06-29 3c05946
Prototype working FireSource.
